Crate libil2cpp

Wrappers and raw bindings for Unity’s libil2cpp

Re-exports

pub use raw::unbox;

pub use raw::value_box_alloc;

pub use raw::WrapRaw;

Modules

raw

Raw il2cpp types and functions

Macros

unsafe_impl_reference_type

Implements the Type trait for a Rust type that is equivalent to a C# reference type

unsafe_impl_value_type

Implements the Type trait for a Rust type that is equivalent to a C# value type

Structs

ByRef

TODO: Add lifetime param Wrapper type which implies the type is ByRef managed lifetime

FieldInfo

Information about a C# field

Gc

Wrapper type which implies the type is GC managed lifetime

Il2CppArray

An il2cpp array

Il2CppClass

An il2cpp class

Il2CppException

An il2cpp exception

Il2CppObject

An il2cpp object

Il2CppReflectionMethod

Object used for reflection of methods

Il2CppReflectionType

Object used for reflection of types

Il2CppString

An il2cpp string

Il2CppType

An il2cpp type

MethodInfo

Information about a C# method

ParameterInfo

Information about a C# parameter

ValueTypePadding

Padding type for value types

Enums

Builtin

Builtin C# types

FindMethodError

Possible errors when looking up a method

Traits

Argument

Trait implemented by types that can be used as C# method arguments

Arguments

Trait implemented by types that can be used as a collection of C# method arguments

Generics

Trait implemented for Rust types which can represent a list of C# generic arguments

ObjectExt

Helper trait for reference types which can be dereferenced to an object

ObjectType

Trait for types that can be treated as Il2CppObject Useful for generic constraints and supporting multiple object-like types

Parameter

Trait implemented by types that can be used as C# method parameters

Parameters

Trait implemented by types that can be used as a collection of C# method parameters

Return

Trait implemented by types that can be used as return types for C# methods

Returned

Trait implemented by types that can be used as return types from C# methods

ThisArgument

Trait implemented by types that can be used as a C# this arguments

ThisParameter

Trait implemented by types that can be used as C# this method parameters

Type

Trait implemented by Rust types that are also C# types

ValueType

Extension trait for value types providing additional functionality

Type Aliases

ByRefMut

Mutable ByRef type alias

Result

Result type for method invocations

Void

Alias for () to represent void return type

Trait Aliases

GcType

Trait alias for types that can be used with the Gc wrapper.