libil2cpp

Module raw

Raw il2cpp types and functions

This module contains raw C types defined in libil2cpp and raw C functions dynamically loaded from libil2cpp.so.

Structs

EventInfo

FieldInfo

GcFunctions

Boehm GC functions used by libil2cpp. gc_alloc_fixed is None if it couldn’t be resolved (only possible on unity2018/il2cpp_v24 without the xref feature).

Il2CppAppContext

Il2CppAppDomain

Il2CppAppDomainSetup

Il2CppArgumentException

Il2CppArray

Il2CppArrayBounds

Il2CppArraySize

Il2CppArrayType

Il2CppAssembly

Il2CppAssemblyName

Il2CppAsyncCall

Il2CppAsyncResult

Il2CppByReference

Il2CppCalendarData

Il2CppCatchPoint

Il2CppClass

Il2CppCodeGenModule

Il2CppCodeGenOptions

Il2CppCodeRegistration

Il2CppCultureData

Il2CppCultureInfo

Il2CppCustomAttrInfo

Il2CppDateTimeFormatInfo

Il2CppDebuggerMetadataRegistration

Il2CppDebuggerTransport

Il2CppDecimal

Il2CppDecimal__bindgen_ty_1__bindgen_ty_1

Il2CppDecimal__bindgen_ty_2__bindgen_ty_1

Il2CppDefaults

Il2CppDelegate

Il2CppDomain

Il2CppDouble

Il2CppErrorWrapper

Il2CppException

Il2CppExceptionWrapper

Il2CppFileTime

Il2CppGenericClass

Il2CppGenericContext

Il2CppGenericInst

Il2CppGenericMethod

Il2CppGenericMethodFunctionsDefinitions

Il2CppGenericMethodIndices

Il2CppGenericParameterInfo

Il2CppGuid

Il2CppHStringHeader

Il2CppHString__

Il2CppIOAsyncResult

Il2CppIOSelectorJob

Il2CppIUnknown

Il2CppImage

Il2CppInterfaceOffsetInfo

Il2CppInternalThread

Il2CppInteropData

Il2CppLongLivedThreadData

Il2CppManagedMemorySnapshot

Il2CppManifestResourceInfo

Il2CppMarshalByRefObject

Il2CppMemoryCallbacks

Il2CppMetadataEventInfo

Il2CppMetadataFieldInfo

Il2CppMetadataMethodInfo

Il2CppMetadataParameterInfo

Il2CppMetadataPropertyInfo

Il2CppMetadataRegistration

Il2CppMethodDebugInfo

Il2CppMethodExecutionContextInfo

Il2CppMethodExecutionContextInfoIndex

Il2CppMethodHeaderInfo

Il2CppMethodInfo

Il2CppMethodMessage

Il2CppMethodScope

Il2CppMethodSpec

Il2CppMulticastDelegate

Il2CppNameToTypeHandleHashTable

Il2CppNumberFormatInfo

Il2CppObject

Il2CppPerfCounters

Il2CppProfiler

Il2CppPropertyInfo

Il2CppRGCTXConstrainedData

Il2CppRGCTXDefinition

Il2CppRange

Il2CppRefCount

Il2CppReflectionAssembly

Il2CppReflectionAssemblyName

Il2CppReflectionEvent

Il2CppReflectionField

Il2CppReflectionGenericMethod

Il2CppReflectionMarshal

Il2CppReflectionMethod

Il2CppReflectionModule

Il2CppReflectionMonoEvent

Il2CppReflectionMonoEventInfo

Il2CppReflectionMonoType

Il2CppReflectionParameter

Il2CppReflectionPointer

Il2CppReflectionProperty

Il2CppReflectionRuntimeType

Il2CppReflectionType

Il2CppRegionInfo

Il2CppRuntimeInterfaceOffsetPair

Il2CppSafeArray

Il2CppSafeArrayBound

Il2CppSafeHandle

Il2CppSequencePoint

Il2CppSequencePointSourceFile

Il2CppSingle

Il2CppSocketAddress

Il2CppSocketAsyncResult

Corresponds to Mono’s internal System.Net.Sockets.Socket.SocketAsyncResult class. Has no relation to Il2CppAsyncResult.

Il2CppSortKey

Il2CppStackFrame

Il2CppStackFrameInfo

Il2CppStatStg

Il2CppString

Il2CppStringBuilder

Il2CppSystemException

Il2CppThread

Il2CppThreadName

Il2CppTokenAdjustorThunkPair

Il2CppTokenIndexMethodTuple

Il2CppTokenRangePair

Il2CppType

Il2CppTypeDefinitionSizes

Il2CppTypeSourceFilePair

Il2CppTypedRef

Il2CppVariant

Il2CppVariant__bindgen_ty_1___tagVARIANT

Il2CppVariant__bindgen_ty_1___tagVARIANT__bindgen_ty_1___tagBRECORD

Il2CppWaitHandle

Il2CppWin32Decimal

Il2CppWin32Decimal__bindgen_ty_1__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_2__bindgen_ty_1

Il2CppWindowsRuntimeFactoryTableEntry

Il2CppWindowsRuntimeTypeName

MemberInfo

MethodInfo

MonitorData

NumberFormatEntryManaged

PropertyInfo

VirtualInvokeData

__BindgenBitfieldUnit

__IncompleteArrayField

___Il2CppMetadataCustomAttributeHandle

___Il2CppMetadataGenericContainerHandle

___Il2CppMetadataGenericParameterHandle

___Il2CppMetadataImageHandle

___Il2CppMetadataMethodHandle

___Il2CppMetadataTypeHandle

__fsid_t

__locale_data

__locale_struct

max_align_t

Constants

ASSEMBLYREF_DISABLEJITCOMPILE_OPTIMIZER_FLAG

ASSEMBLYREF_ENABLEJITCOMPILE_TRACKING_FLAG

ASSEMBLYREF_FULL_PUBLIC_KEY_FLAG

ASSEMBLYREF_RETARGETABLE_FLAG

FIELD_ATTRIBUTE_ASSEMBLY

FIELD_ATTRIBUTE_COMPILER_CONTROLLED

FIELD_ATTRIBUTE_FAMILY

FIELD_ATTRIBUTE_FAM_AND_ASSEM

FIELD_ATTRIBUTE_FAM_OR_ASSEM

FIELD_ATTRIBUTE_FIELD_ACCESS_MASK

FIELD_ATTRIBUTE_HAS_DEFAULT

FIELD_ATTRIBUTE_HAS_FIELD_MARSHAL

FIELD_ATTRIBUTE_HAS_FIELD_RVA

FIELD_ATTRIBUTE_INIT_ONLY

FIELD_ATTRIBUTE_LITERAL

FIELD_ATTRIBUTE_NOT_SERIALIZED

FIELD_ATTRIBUTE_PINVOKE_IMPL

FIELD_ATTRIBUTE_PRIVATE

FIELD_ATTRIBUTE_PUBLIC

FIELD_ATTRIBUTE_RESERVED_MASK

FIELD_ATTRIBUTE_RT_SPECIAL_NAME

FIELD_ATTRIBUTE_SPECIAL_NAME

FIELD_ATTRIBUTE_STATIC

FIXME

IL2CPP_API_DYNAMIC_NO_DLSYM

IL2CPP_BIG_ENDIAN

IL2CPP_BINARY

IL2CPP_BINARY_SECTION_NAME

IL2CPP_BYTE_ORDER

IL2CPP_CAN_USE_MULTIPLE_SYMBOL_MAPS

IL2CPP_CXX_ABI_MSVC

IL2CPP_DEBUG

IL2CPP_DEVELOPMENT

IL2CPP_DIR_SEPARATOR

IL2CPP_DISABLE_FULL_MESSAGES

IL2CPP_ENABLE_DEFERRED_GC

IL2CPP_ENABLE_MONO_BUG_EMULATION

IL2CPP_GC_BOEHM

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_CONTRAVARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_COVARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_DEFAULT_CONSTRUCTOR_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_NON_VARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_NOT_NULLABLE_VALUE_TYPE_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_REFERENCE_TYPE_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_SPECIAL_CONSTRAINT_MASK

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_VARIANCE_MASK

IL2CPP_HAS_DELETED_FUNCTIONS

IL2CPP_HAS_GC_DESCRIPTORS

IL2CPP_LITTLE_ENDIAN

IL2CPP_PAGE_SIZE

IL2CPP_PLATFORM_DISABLE_LIBC_PINVOKE

IL2CPP_PLATFORM_OVERRIDES_STD_FILE_HANDLES

IL2CPP_PLATFORM_SUPPORTS_CPU_INFO

IL2CPP_PLATFORM_SUPPORTS_DEBUGGER_PRESENT

IL2CPP_PLATFORM_SUPPORTS_SYSTEM_CERTIFICATES

IL2CPP_PLATFORM_SUPPORTS_TIMEZONEINFO

IL2CPP_POINTER_SPARE_BITS

IL2CPP_SIZEOF_STRUCT_WITH_NO_INSTANCE_FIELDS

IL2CPP_SIZEOF_VOID_P

IL2CPP_SUPPORTS_BROKERED_FILESYSTEM

IL2CPP_SUPPORTS_CONSOLE_EXTENSION

IL2CPP_SUPPORT_SOCKETS_POSIX_API

IL2CPP_TARGET_ANDROID

IL2CPP_TARGET_ARM64

IL2CPP_TARGET_ARM64E

IL2CPP_TARGET_ARMV7

IL2CPP_TARGET_DARWIN

IL2CPP_TARGET_EMBEDDED_LINUX

IL2CPP_TARGET_GAMECORE_XBOX

IL2CPP_TARGET_IOS

IL2CPP_TARGET_JAVASCRIPT

IL2CPP_TARGET_LINUX

IL2CPP_TARGET_N3DS

IL2CPP_TARGET_OSX

IL2CPP_TARGET_PS4

IL2CPP_TARGET_PSP2

IL2CPP_TARGET_QNX

IL2CPP_TARGET_SWITCH

IL2CPP_TARGET_WINDOWS

IL2CPP_TARGET_WINDOWS_DESKTOP

IL2CPP_TARGET_WINDOWS_GAMES

IL2CPP_TARGET_WINRT

IL2CPP_TARGET_X64

IL2CPP_TARGET_X86

IL2CPP_TARGET_XBOXONE

IL2CPP_THREADS_STD

IL2CPP_THREAD_IMPL_HAS_COM_APARTMENTS

IL2CPP_USE_BASELIB_FAST_READER_RWL

IL2CPP_USE_GENERIC_CPU_INFO

IL2CPP_USE_GENERIC_PROCESS

IL2CPP_USE_NETWORK_ACCESS_HANDLER

IL2CPP_USE_SEND_NOSIGNAL

IL2CPP_USE_STD_THREAD

IL2CPP_VALIDATE_FIELD_LAYOUT

IL2CPP_ZERO_LEN_ARRAY

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

Il2CppCallConvention_IL2CPP_CALL_C

Il2CppCallConvention_IL2CPP_CALL_DEFAULT

Il2CppCallConvention_IL2CPP_CALL_FASTCALL

Il2CppCallConvention_IL2CPP_CALL_STDCALL

Il2CppCallConvention_IL2CPP_CALL_THISCALL

Il2CppCallConvention_IL2CPP_CALL_VARARG

Il2CppCallType_Il2Cpp_CallType_BeginInvoke

Il2CppCallType_Il2Cpp_CallType_EndInvoke

Il2CppCallType_Il2Cpp_CallType_OneWay

Il2CppCallType_Il2Cpp_CallType_Sync

Il2CppCharSet_CHARSET_ANSI

Il2CppCharSet_CHARSET_NOT_SPECIFIED

Il2CppCharSet_CHARSET_UNICODE

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_EQ

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_GT

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_LT

Il2CppGCEvent_IL2CPP_GC_EVENT_END

Il2CppGCEvent_IL2CPP_GC_EVENT_MARK_END

Il2CppGCEvent_IL2CPP_GC_EVENT_MARK_START

Il2CppGCEvent_IL2CPP_GC_EVENT_POST_START_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_POST_STOP_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_PRE_START_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_PRE_STOP_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_RECLAIM_END

Il2CppGCEvent_IL2CPP_GC_EVENT_RECLAIM_START

Il2CppGCEvent_IL2CPP_GC_EVENT_START

Il2CppGCMode_IL2CPP_GC_MODE_DISABLED

Il2CppGCMode_IL2CPP_GC_MODE_ENABLED

Il2CppGCMode_IL2CPP_GC_MODE_MANUAL

Il2CppProfileFileIOKind_IL2CPP_PROFILE_FILEIO_READ

Il2CppProfileFileIOKind_IL2CPP_PROFILE_FILEIO_WRITE

Il2CppProfileFlags_IL2CPP_PROFILE_ALLOCATIONS

Il2CppProfileFlags_IL2CPP_PROFILE_APPDOMAIN_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_ASSEMBLY_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_CLASS_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_COVERAGE

Il2CppProfileFlags_IL2CPP_PROFILE_ENTER_LEAVE

Il2CppProfileFlags_IL2CPP_PROFILE_EXCEPTIONS

Il2CppProfileFlags_IL2CPP_PROFILE_FILEIO

Il2CppProfileFlags_IL2CPP_PROFILE_GC

Il2CppProfileFlags_IL2CPP_PROFILE_GC_MOVES

Il2CppProfileFlags_IL2CPP_PROFILE_INLINING

Il2CppProfileFlags_IL2CPP_PROFILE_INS_COVERAGE

Il2CppProfileFlags_IL2CPP_PROFILE_IOMAP_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_JIT_COMPILATION

Il2CppProfileFlags_IL2CPP_PROFILE_METHOD_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_MODULE_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_MONITOR_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_NONE

Il2CppProfileFlags_IL2CPP_PROFILE_REMOTING

Il2CppProfileFlags_IL2CPP_PROFILE_STATISTICAL

Il2CppProfileFlags_IL2CPP_PROFILE_THREADS

Il2CppProfileFlags_IL2CPP_PROFILE_TRANSITIONS

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_ARRAY

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_CLASS

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_CONSTRAINED

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_INVALID

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_METHOD

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_TYPE

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_ANOTHER_ASSEMBLY

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_EMBEDDED

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_IN_MANIFEST

Il2CppRuntimeUnhandledExceptionPolicy_IL2CPP_UNHANDLED_POLICY_CURRENT

Il2CppRuntimeUnhandledExceptionPolicy_IL2CPP_UNHANDLED_POLICY_LEGACY

Il2CppStat_IL2CPP_STAT_CLASS_STATIC_DATA_SIZE

Il2CppStat_IL2CPP_STAT_GENERIC_CLASS_COUNT

Il2CppStat_IL2CPP_STAT_GENERIC_INSTANCE_COUNT

Il2CppStat_IL2CPP_STAT_INFLATED_METHOD_COUNT

Il2CppStat_IL2CPP_STAT_INFLATED_TYPE_COUNT

Il2CppStat_IL2CPP_STAT_INITIALIZED_CLASS_COUNT

Il2CppStat_IL2CPP_STAT_METHOD_COUNT

Il2CppStat_IL2CPP_STAT_NEW_OBJECT_COUNT

Il2CppTokenType_IL2CPP_TOKEN_ASSEMBLY

Il2CppTokenType_IL2CPP_TOKEN_ASSEMBLY_REF

Il2CppTokenType_IL2CPP_TOKEN_CUSTOM_ATTRIBUTE

Il2CppTokenType_IL2CPP_TOKEN_EVENT

Il2CppTokenType_IL2CPP_TOKEN_EXPORTED_TYPE

Il2CppTokenType_IL2CPP_TOKEN_FIELD_DEF

Il2CppTokenType_IL2CPP_TOKEN_FILE

Il2CppTokenType_IL2CPP_TOKEN_GENERIC_PARAM

Il2CppTokenType_IL2CPP_TOKEN_INTERFACE_IMPL

Il2CppTokenType_IL2CPP_TOKEN_MANIFEST_RESOURCE

Il2CppTokenType_IL2CPP_TOKEN_MEMBER_REF

Il2CppTokenType_IL2CPP_TOKEN_METHOD_DEF

Il2CppTokenType_IL2CPP_TOKEN_METHOD_SPEC

Il2CppTokenType_IL2CPP_TOKEN_MODULE

Il2CppTokenType_IL2CPP_TOKEN_MODULE_REF

Il2CppTokenType_IL2CPP_TOKEN_PARAM_DEF

Il2CppTokenType_IL2CPP_TOKEN_PERMISSION

Il2CppTokenType_IL2CPP_TOKEN_PROPERTY

Il2CppTokenType_IL2CPP_TOKEN_SIGNATURE

Il2CppTokenType_IL2CPP_TOKEN_TYPE_DEF

Il2CppTokenType_IL2CPP_TOKEN_TYPE_REF

Il2CppTokenType_IL2CPP_TOKEN_TYPE_SPEC

Il2CppTypeEnum_IL2CPP_TYPE_ARRAY

Il2CppTypeEnum_IL2CPP_TYPE_BOOLEAN

Il2CppTypeEnum_IL2CPP_TYPE_BYREF

Il2CppTypeEnum_IL2CPP_TYPE_CHAR

Il2CppTypeEnum_IL2CPP_TYPE_CLASS

Il2CppTypeEnum_IL2CPP_TYPE_CMOD_OPT

Il2CppTypeEnum_IL2CPP_TYPE_CMOD_REQD

Il2CppTypeEnum_IL2CPP_TYPE_END

Il2CppTypeEnum_IL2CPP_TYPE_ENUM

Il2CppTypeEnum_IL2CPP_TYPE_FNPTR

Il2CppTypeEnum_IL2CPP_TYPE_GENERICINST

Il2CppTypeEnum_IL2CPP_TYPE_I

Il2CppTypeEnum_IL2CPP_TYPE_I1

Il2CppTypeEnum_IL2CPP_TYPE_I2

Il2CppTypeEnum_IL2CPP_TYPE_I4

Il2CppTypeEnum_IL2CPP_TYPE_I8

Il2CppTypeEnum_IL2CPP_TYPE_IL2CPP_TYPE_INDEX

Il2CppTypeEnum_IL2CPP_TYPE_INTERNAL

Il2CppTypeEnum_IL2CPP_TYPE_MODIFIER

Il2CppTypeEnum_IL2CPP_TYPE_MVAR

Il2CppTypeEnum_IL2CPP_TYPE_OBJECT

Il2CppTypeEnum_IL2CPP_TYPE_PINNED

Il2CppTypeEnum_IL2CPP_TYPE_PTR

Il2CppTypeEnum_IL2CPP_TYPE_R4

Il2CppTypeEnum_IL2CPP_TYPE_R8

Il2CppTypeEnum_IL2CPP_TYPE_SENTINEL

Il2CppTypeEnum_IL2CPP_TYPE_STRING

Il2CppTypeEnum_IL2CPP_TYPE_SZARRAY

Il2CppTypeEnum_IL2CPP_TYPE_TYPEDBYREF

Il2CppTypeEnum_IL2CPP_TYPE_U

Il2CppTypeEnum_IL2CPP_TYPE_U1

Il2CppTypeEnum_IL2CPP_TYPE_U2

Il2CppTypeEnum_IL2CPP_TYPE_U4

Il2CppTypeEnum_IL2CPP_TYPE_U8

Il2CppTypeEnum_IL2CPP_TYPE_VALUETYPE

Il2CppTypeEnum_IL2CPP_TYPE_VAR

Il2CppTypeEnum_IL2CPP_TYPE_VOID

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_ASSEMBLY_QUALIFIED

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_FULL_NAME

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_IL

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_REFLECTION

Il2CppVarType_IL2CPP_VT_ARRAY

Il2CppVarType_IL2CPP_VT_BLOB

Il2CppVarType_IL2CPP_VT_BLOB_OBJECT

Il2CppVarType_IL2CPP_VT_BOOL

Il2CppVarType_IL2CPP_VT_BSTR

Il2CppVarType_IL2CPP_VT_BSTR_BLOB

Il2CppVarType_IL2CPP_VT_BYREF

Il2CppVarType_IL2CPP_VT_CARRAY

Il2CppVarType_IL2CPP_VT_CF

Il2CppVarType_IL2CPP_VT_CLSID

Il2CppVarType_IL2CPP_VT_CY

Il2CppVarType_IL2CPP_VT_DATE

Il2CppVarType_IL2CPP_VT_DECIMAL

Il2CppVarType_IL2CPP_VT_DISPATCH

Il2CppVarType_IL2CPP_VT_EMPTY

Il2CppVarType_IL2CPP_VT_ERROR

Il2CppVarType_IL2CPP_VT_FILETIME

Il2CppVarType_IL2CPP_VT_HRESULT

Il2CppVarType_IL2CPP_VT_I1

Il2CppVarType_IL2CPP_VT_I2

Il2CppVarType_IL2CPP_VT_I4

Il2CppVarType_IL2CPP_VT_I8

Il2CppVarType_IL2CPP_VT_ILLEGAL

Il2CppVarType_IL2CPP_VT_ILLEGALMASKED

Il2CppVarType_IL2CPP_VT_INT

Il2CppVarType_IL2CPP_VT_INT_PTR

Il2CppVarType_IL2CPP_VT_LPSTR

Il2CppVarType_IL2CPP_VT_LPWSTR

Il2CppVarType_IL2CPP_VT_NULL

Il2CppVarType_IL2CPP_VT_PTR

Il2CppVarType_IL2CPP_VT_R4

Il2CppVarType_IL2CPP_VT_R8

Il2CppVarType_IL2CPP_VT_RECORD

Il2CppVarType_IL2CPP_VT_RESERVED

Il2CppVarType_IL2CPP_VT_SAFEARRAY

Il2CppVarType_IL2CPP_VT_STORAGE

Il2CppVarType_IL2CPP_VT_STORED_OBJECT

Il2CppVarType_IL2CPP_VT_STREAM

Il2CppVarType_IL2CPP_VT_STREAMED_OBJECT

Il2CppVarType_IL2CPP_VT_TYPEMASK

Il2CppVarType_IL2CPP_VT_UI1

Il2CppVarType_IL2CPP_VT_UI2

Il2CppVarType_IL2CPP_VT_UI4

Il2CppVarType_IL2CPP_VT_UI8

Il2CppVarType_IL2CPP_VT_UINT

Il2CppVarType_IL2CPP_VT_UINT_PTR

Il2CppVarType_IL2CPP_VT_UNKNOWN

Il2CppVarType_IL2CPP_VT_USERDEFINED

Il2CppVarType_IL2CPP_VT_VARIANT

Il2CppVarType_IL2CPP_VT_VECTOR

Il2CppVarType_IL2CPP_VT_VERSIONED_STREAM

Il2CppVarType_IL2CPP_VT_VOID

Il2CppWindowsRuntimeTypeKind_kTypeKindCustom

Il2CppWindowsRuntimeTypeKind_kTypeKindMetadata

Il2CppWindowsRuntimeTypeKind_kTypeKindPrimitive

MAXIMUM_NESTED_GENERICS_EXCEPTION_MESSAGE

METHOD_ATTRIBUTE_ABSTRACT

METHOD_ATTRIBUTE_ASSEM

METHOD_ATTRIBUTE_COMPILER_CONTROLLED

METHOD_ATTRIBUTE_FAMILY

METHOD_ATTRIBUTE_FAM_AND_ASSEM

METHOD_ATTRIBUTE_FAM_OR_ASSEM

METHOD_ATTRIBUTE_FINAL

METHOD_ATTRIBUTE_HAS_SECURITY

METHOD_ATTRIBUTE_HIDE_BY_SIG

METHOD_ATTRIBUTE_MEMBER_ACCESS_MASK

METHOD_ATTRIBUTE_NEW_SLOT

METHOD_ATTRIBUTE_PINVOKE_IMPL

METHOD_ATTRIBUTE_PRIVATE

METHOD_ATTRIBUTE_PUBLIC

METHOD_ATTRIBUTE_REQUIRE_SEC_OBJECT

METHOD_ATTRIBUTE_RESERVED_MASK

METHOD_ATTRIBUTE_REUSE_SLOT

METHOD_ATTRIBUTE_RT_SPECIAL_NAME

METHOD_ATTRIBUTE_SPECIAL_NAME

METHOD_ATTRIBUTE_STATIC

METHOD_ATTRIBUTE_STRICT

METHOD_ATTRIBUTE_UNMANAGED_EXPORT

METHOD_ATTRIBUTE_VIRTUAL

METHOD_ATTRIBUTE_VTABLE_LAYOUT_MASK

METHOD_IMPL_ATTRIBUTE_CODE_TYPE_MASK

METHOD_IMPL_ATTRIBUTE_FORWARD_REF

METHOD_IMPL_ATTRIBUTE_IL

METHOD_IMPL_ATTRIBUTE_INTERNAL_CALL

METHOD_IMPL_ATTRIBUTE_MANAGED

METHOD_IMPL_ATTRIBUTE_MANAGED_MASK

METHOD_IMPL_ATTRIBUTE_MAX_METHOD_IMPL_VAL

METHOD_IMPL_ATTRIBUTE_NATIVE

METHOD_IMPL_ATTRIBUTE_NOINLINING

METHOD_IMPL_ATTRIBUTE_OPTIL

METHOD_IMPL_ATTRIBUTE_PRESERVE_SIG

METHOD_IMPL_ATTRIBUTE_RUNTIME

METHOD_IMPL_ATTRIBUTE_SYNCHRONIZED

METHOD_IMPL_ATTRIBUTE_UNMANAGED

MethodVariableKind_kMethodVariableKind_LocalVariable

MethodVariableKind_kMethodVariableKind_Parameter

MethodVariableKind_kMethodVariableKind_This

NEED_TO_ZERO_PTRFREE

PARAM_ATTRIBUTE_HAS_DEFAULT

PARAM_ATTRIBUTE_HAS_FIELD_MARSHAL

PARAM_ATTRIBUTE_IN

PARAM_ATTRIBUTE_OPTIONAL

PARAM_ATTRIBUTE_OUT

PARAM_ATTRIBUTE_RESERVED_MASK

PARAM_ATTRIBUTE_UNUSED

PTRDIFF_MAX

PTRDIFF_MIN

PUBLIC_KEY_BYTE_LENGTH

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

SequencePointKind_kSequencePointKind_Normal

SequencePointKind_kSequencePointKind_StepOut

THREAD_STATIC_FIELD_OFFSET

TYPE_ATTRIBUTE_ABSTRACT

TYPE_ATTRIBUTE_ANSI_CLASS

TYPE_ATTRIBUTE_AUTO_CLASS

TYPE_ATTRIBUTE_AUTO_LAYOUT

TYPE_ATTRIBUTE_BEFORE_FIELD_INIT

TYPE_ATTRIBUTE_CLASS

TYPE_ATTRIBUTE_CLASS_SEMANTIC_MASK

TYPE_ATTRIBUTE_EXPLICIT_LAYOUT

TYPE_ATTRIBUTE_FORWARDER

TYPE_ATTRIBUTE_HAS_SECURITY

TYPE_ATTRIBUTE_IMPORT

TYPE_ATTRIBUTE_INTERFACE

TYPE_ATTRIBUTE_LAYOUT_MASK

TYPE_ATTRIBUTE_NESTED_ASSEMBLY

TYPE_ATTRIBUTE_NESTED_FAMILY

TYPE_ATTRIBUTE_NESTED_FAM_AND_ASSEM

TYPE_ATTRIBUTE_NESTED_FAM_OR_ASSEM

TYPE_ATTRIBUTE_NESTED_PRIVATE

TYPE_ATTRIBUTE_NESTED_PUBLIC

TYPE_ATTRIBUTE_NOT_PUBLIC

TYPE_ATTRIBUTE_PUBLIC

TYPE_ATTRIBUTE_RESERVED_MASK

TYPE_ATTRIBUTE_RT_SPECIAL_NAME

TYPE_ATTRIBUTE_SEALED

TYPE_ATTRIBUTE_SEQUENTIAL_LAYOUT

TYPE_ATTRIBUTE_SERIALIZABLE

TYPE_ATTRIBUTE_SPECIAL_NAME

TYPE_ATTRIBUTE_STRING_FORMAT_MASK

TYPE_ATTRIBUTE_UNICODE_CLASS

TYPE_ATTRIBUTE_VISIBILITY_MASK

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_TYPES_LOCALE_T_H

_BITS_TYPES___LOCALE_T_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_STRINGS_H

_STRING_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_C2X_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__glibc_c99_flexarr_available

ipv6AddressSize

kCustomAttributeIndexInvalid

kDefaultValueIndexNull

kEventIndexInvalid

kFieldIndexInvalid

kGenericContainerIndexInvalid

kGenericParameterIndexInvalid

kIl2CppInt32Max

kIl2CppInt32Min

kIl2CppInt64Max

kIl2CppInt64Min

kIl2CppIntPtrMax

kIl2CppIntPtrMin

kIl2CppOffsetOfArrayBounds

kIl2CppOffsetOfArrayLength

kIl2CppSizeOfArray

kIl2CppUInt32Max

kInteropDataIndexInvalid

kInvalidIl2CppMethodSlot

kMethodIndexInvalid

kPropertyIndexInvalid

kPublicKeyByteLength

kRGCTXIndexInvalid

kStringLiteralIndexInvalid

kTypeDefinitionIndexInvalid

kTypeIndexInvalid

Statics

LIBIL2CPP

il2cpp_defaults⚠

kIl2CppNewLine⚠

kIl2CppUInt64Max⚠

kIl2CppUIntPtrMax⚠

Traits

WrapRaw

Safe wrapper around a raw il2cpp type which can be used in its place

Functions

__memcmpeq⚠

__mempcpy⚠

__stpcpy⚠

__stpncpy⚠

__strtok_r⚠

array_class_get⚠

array_new⚠

assembly_get_image⚠

bcmp⚠

bcopy⚠

bzero⚠

class_from_il2cpp_type⚠

class_from_name⚠

class_from_system_type⚠

class_get_method_from_name⚠

class_get_type⚠

class_is_assignable_from⚠

domain_get⚠

domain_get_assemblies⚠

explicit_bzero⚠

fake_value_box⚠

Boxes a value type into an Il2CppObject without allocating or copying, mirroring beatsaber-hook’s to_object<Box = true>(fake_box = true)

ffs⚠

ffsl⚠

ffsll⚠

field_get_value⚠

field_set_value⚠

gc_alloc_fixed⚠

gc_free_fixed⚠

il2cpp_add_internal_call⚠

il2cpp_alloc⚠

il2cpp_allocation_granularity⚠

il2cpp_array_class_get⚠

il2cpp_array_element_size⚠

il2cpp_array_get_byte_length⚠

il2cpp_array_length⚠

il2cpp_array_new⚠

il2cpp_array_new_full⚠

il2cpp_array_new_specific⚠

il2cpp_array_object_header_size⚠

il2cpp_assembly_get_image⚠

il2cpp_bounded_array_class_get⚠

il2cpp_capture_memory_snapshot⚠

il2cpp_class_array_element_size⚠

il2cpp_class_enum_basetype⚠

il2cpp_class_for_each⚠

il2cpp_class_from_il2cpp_type⚠

il2cpp_class_from_name⚠

il2cpp_class_from_system_type⚠

il2cpp_class_from_type⚠

il2cpp_class_get_assemblyname⚠

il2cpp_class_get_bitmap⚠

il2cpp_class_get_bitmap_size⚠

il2cpp_class_get_data_size⚠

il2cpp_class_get_declaring_type⚠

il2cpp_class_get_element_class⚠

il2cpp_class_get_events⚠

il2cpp_class_get_field_from_name⚠

il2cpp_class_get_fields⚠

il2cpp_class_get_flags⚠

il2cpp_class_get_image⚠

il2cpp_class_get_interfaces⚠

il2cpp_class_get_method_from_name⚠

il2cpp_class_get_methods⚠

il2cpp_class_get_name⚠

il2cpp_class_get_namespace⚠

il2cpp_class_get_nested_types⚠

il2cpp_class_get_parent⚠

il2cpp_class_get_properties⚠

il2cpp_class_get_property_from_name⚠

il2cpp_class_get_rank⚠

il2cpp_class_get_static_field_data⚠

il2cpp_class_get_type⚠

il2cpp_class_get_type_token⚠

il2cpp_class_get_userdata_offset⚠

il2cpp_class_has_attribute⚠

il2cpp_class_has_parent⚠

il2cpp_class_has_references⚠

il2cpp_class_instance_size⚠

il2cpp_class_is_abstract⚠

il2cpp_class_is_assignable_from⚠

il2cpp_class_is_blittable⚠

il2cpp_class_is_enum⚠

il2cpp_class_is_generic⚠

il2cpp_class_is_inflated⚠

il2cpp_class_is_inited⚠

il2cpp_class_is_interface⚠

il2cpp_class_is_subclass_of⚠

il2cpp_class_is_valuetype⚠

il2cpp_class_num_fields⚠

il2cpp_class_set_userdata⚠

il2cpp_class_value_size⚠

il2cpp_current_thread_get_frame_at⚠

il2cpp_current_thread_get_stack_depth⚠

il2cpp_current_thread_get_top_frame⚠

il2cpp_current_thread_walk_frame_stack⚠

il2cpp_custom_attrs_construct⚠

il2cpp_custom_attrs_free⚠

il2cpp_custom_attrs_from_class⚠

il2cpp_custom_attrs_from_field⚠

il2cpp_custom_attrs_from_method⚠

il2cpp_custom_attrs_get_attr⚠

il2cpp_custom_attrs_has_attr⚠

il2cpp_debug_get_method_info⚠

il2cpp_debugger_set_agent_options⚠

il2cpp_domain_assembly_open⚠

il2cpp_domain_get⚠

il2cpp_domain_get_assemblies⚠

il2cpp_exception_from_name_msg⚠

il2cpp_field_get_flags⚠

il2cpp_field_get_name⚠

il2cpp_field_get_offset⚠

il2cpp_field_get_parent⚠

il2cpp_field_get_type⚠

il2cpp_field_get_value⚠

il2cpp_field_get_value_object⚠

il2cpp_field_has_attribute⚠

il2cpp_field_is_literal⚠

il2cpp_field_set_value⚠

il2cpp_field_set_value_object⚠

il2cpp_field_static_get_value⚠

il2cpp_field_static_set_value⚠

il2cpp_format_exception⚠

il2cpp_format_stack_trace⚠

il2cpp_free⚠

il2cpp_free_captured_memory_snapshot⚠

il2cpp_gc_alloc_fixed⚠

il2cpp_gc_collect⚠

il2cpp_gc_collect_a_little⚠

il2cpp_gc_disable⚠

il2cpp_gc_enable⚠

il2cpp_gc_foreach_heap⚠

il2cpp_gc_free_fixed⚠

il2cpp_gc_get_heap_size⚠

il2cpp_gc_get_max_time_slice_ns⚠

il2cpp_gc_get_used_size⚠

il2cpp_gc_has_strict_wbarriers⚠

il2cpp_gc_is_disabled⚠

il2cpp_gc_is_incremental⚠

il2cpp_gc_set_external_allocation_tracker⚠

il2cpp_gc_set_external_wbarrier_tracker⚠

il2cpp_gc_set_max_time_slice_ns⚠

il2cpp_gc_set_mode⚠

il2cpp_gc_start_incremental_collection⚠

il2cpp_gc_wbarrier_set_field⚠

il2cpp_gchandle_foreach_get_target⚠

il2cpp_gchandle_free⚠

il2cpp_gchandle_get_target⚠

il2cpp_gchandle_new⚠

il2cpp_gchandle_new_weakref⚠

il2cpp_get_corlib⚠

il2cpp_get_exception_argument_null⚠

il2cpp_image_get_assembly⚠

il2cpp_image_get_class⚠

il2cpp_image_get_class_count⚠

il2cpp_image_get_entry_point⚠

il2cpp_image_get_filename⚠

il2cpp_image_get_name⚠

il2cpp_init⚠

il2cpp_init_utf16⚠

il2cpp_is_debugger_attached⚠

il2cpp_is_vm_thread⚠

il2cpp_method_get_class⚠

il2cpp_method_get_declaring_type⚠

il2cpp_method_get_flags⚠

il2cpp_method_get_from_reflection⚠

il2cpp_method_get_name⚠

il2cpp_method_get_object⚠

il2cpp_method_get_param⚠

il2cpp_method_get_param_count⚠

il2cpp_method_get_param_name⚠

il2cpp_method_get_return_type⚠

il2cpp_method_get_token⚠

il2cpp_method_has_attribute⚠

il2cpp_method_is_generic⚠

il2cpp_method_is_inflated⚠

il2cpp_method_is_instance⚠

il2cpp_monitor_enter⚠

il2cpp_monitor_exit⚠

il2cpp_monitor_pulse⚠

il2cpp_monitor_pulse_all⚠

il2cpp_monitor_try_enter⚠

il2cpp_monitor_try_wait⚠

il2cpp_monitor_wait⚠

il2cpp_native_stack_trace⚠

il2cpp_object_get_class⚠

il2cpp_object_get_size⚠

il2cpp_object_get_virtual_method⚠

il2cpp_object_header_size⚠

il2cpp_object_new⚠

il2cpp_object_unbox⚠

il2cpp_offset_of_array_bounds_in_array_object_header⚠

il2cpp_offset_of_array_length_in_array_object_header⚠

il2cpp_override_stack_backtrace⚠

il2cpp_profiler_install⚠

il2cpp_profiler_install_allocation⚠

il2cpp_profiler_install_enter_leave⚠

il2cpp_profiler_install_fileio⚠

il2cpp_profiler_install_gc⚠

il2cpp_profiler_install_thread⚠

il2cpp_profiler_set_events⚠

il2cpp_property_get_flags⚠

il2cpp_property_get_get_method⚠

il2cpp_property_get_name⚠

il2cpp_property_get_parent⚠

il2cpp_property_get_set_method⚠

il2cpp_raise_exception⚠

il2cpp_register_debugger_agent_transport⚠

il2cpp_register_log_callback⚠

il2cpp_resolve_icall⚠

il2cpp_runtime_class_init⚠

il2cpp_runtime_invoke⚠

il2cpp_runtime_invoke_convert_args⚠

il2cpp_runtime_object_init⚠

il2cpp_runtime_object_init_exception⚠

il2cpp_runtime_unhandled_exception_policy_set⚠

il2cpp_set_commandline_arguments⚠

il2cpp_set_commandline_arguments_utf16⚠

il2cpp_set_config⚠

il2cpp_set_config_dir⚠

il2cpp_set_config_utf16⚠

il2cpp_set_data_dir⚠

il2cpp_set_default_thread_affinity⚠

il2cpp_set_find_plugin_callback⚠

il2cpp_set_memory_callbacks⚠

il2cpp_set_temp_dir⚠

il2cpp_shutdown⚠

il2cpp_start_gc_world⚠

il2cpp_stats_dump_to_file⚠

il2cpp_stats_get_value⚠

il2cpp_stop_gc_world⚠

il2cpp_string_chars⚠

il2cpp_string_intern⚠

il2cpp_string_is_interned⚠

il2cpp_string_length⚠

il2cpp_string_new⚠

il2cpp_string_new_len⚠

il2cpp_string_new_utf16⚠

il2cpp_string_new_wrapper⚠

il2cpp_thread_attach⚠

il2cpp_thread_current⚠

il2cpp_thread_detach⚠

il2cpp_thread_get_all_attached_threads⚠

il2cpp_thread_get_frame_at⚠

il2cpp_thread_get_stack_depth⚠

il2cpp_thread_get_top_frame⚠

il2cpp_thread_walk_frame_stack⚠

il2cpp_type_equals⚠

il2cpp_type_get_assembly_qualified_name⚠

il2cpp_type_get_attrs⚠

il2cpp_type_get_class_or_element_class⚠

il2cpp_type_get_name⚠

il2cpp_type_get_name_chunked⚠

il2cpp_type_get_object⚠

il2cpp_type_get_reflection_name⚠

il2cpp_type_get_type⚠

il2cpp_type_is_byref⚠

il2cpp_type_is_pointer_type⚠

il2cpp_type_is_static⚠

il2cpp_unhandled_exception⚠

il2cpp_unity_install_unitytls_interface⚠

il2cpp_unity_liveness_allocate_struct⚠

il2cpp_unity_liveness_calculation_from_root⚠

il2cpp_unity_liveness_calculation_from_statics⚠

il2cpp_unity_liveness_finalize⚠

il2cpp_unity_liveness_free_struct⚠

il2cpp_unity_set_android_network_up_state_func⚠

il2cpp_value_box⚠

index⚠

memccpy⚠

memchr⚠

memcmp⚠

memcpy⚠

memmem⚠

memmove⚠

mempcpy⚠

memset⚠

method_get_from_reflection⚠

method_get_object⚠

method_is_generic⚠

object_new⚠

object_unbox⚠

raise_exception⚠

resolve_icall⚠

rindex⚠

runtime_invoke⚠

stpcpy⚠

stpncpy⚠

strcasecmp⚠

strcasecmp_l⚠

strcasestr⚠

strcat⚠

strchr⚠

strchrnul⚠

strcmp⚠

strcoll⚠

strcoll_l⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_l⚠

strerror_r⚠

string_new_len⚠

strlcat⚠

strlcpy⚠

strlen⚠

strncasecmp⚠

strncasecmp_l⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsep⚠

strsignal⚠

strspn⚠

strstr⚠

strtok⚠

strtok_r⚠

strxfrm⚠

strxfrm_l⚠

type_get_name⚠

type_get_object⚠

unbox⚠

Unboxes a value type stored as an Il2CppObject

value_box⚠

value_box_alloc⚠

Boxes a value type into a BoxedValue<T>

Type Aliases

AssemblyExportedTypeIndex

AssemblyIndex

AssemblyTypeIndex

CreateCCWFunc

CultureInfoChangedCallback

CustomAttributeIndex

DefaultValueDataIndex

DefaultValueIndex

EventIndex

FieldIndex

GcAllocFixedFn

GcAllocFixed allocates a fixed-size object in the Boehm GC heap. The object is not movable and will not be collected until it is explicitly freed.

GcFreeFixedFn

GcFreeFixed frees a fixed-size object allocated with GcAllocFixed. The object must not be used after it is freed.

GenericContainerIndex

GenericContainerParameterIndex

GenericInstIndex

GenericMethodIndex

GenericParameterConstraintIndex

GenericParameterIndex

IL2CPP_VARIANT_BOOL

Il2CppAndroidUpStateFunc

Il2CppBacktraceFunc

Il2CppCallConvention

Il2CppCallType

Il2CppChar

Il2CppCharSet

Il2CppDecimalCompareResult

Il2CppFrameWalkFunc

Il2CppFullySharedGenericAny

Il2CppFullySharedGenericStruct

Il2CppGCEvent

Il2CppGCMode

Il2CppHString

Il2CppLogCallback

Il2CppMetadataCustomAttributeHandle

Il2CppMetadataGenericContainerHandle

Il2CppMetadataGenericParameterHandle

Il2CppMetadataImageHandle

Il2CppMetadataMethodDefinitionHandle

Il2CppMetadataTypeHandle

Il2CppMethodPointer

Il2CppMethodSlot

Il2CppNativeChar

Il2CppProfileAllocFunc

Il2CppProfileFileIOFunc

Il2CppProfileFileIOKind

Il2CppProfileFlags

Il2CppProfileFunc

Il2CppProfileGCFunc

Il2CppProfileGCResizeFunc

Il2CppProfileMethodFunc

Il2CppProfileThreadFunc

Il2CppRGCTXDataType

Il2CppResourceLocation

Il2CppRuntimeUnhandledExceptionPolicy

Il2CppSetFindPlugInCallback

Il2CppStat

Il2CppTokenType

Il2CppTypeEnum

Il2CppTypeNameFormat

Il2CppVTable

Il2CppVarType

Il2CppWindowsRuntimeTypeKind

ImageIndex

InterfacesIndex

InteropDataIndex

InvokerMethod

MethodIndex

MethodParameterIndex

MethodVariableKind

NestedTypeIndex

PInvokeMarshalCleanupFunc

PInvokeMarshalFromNativeFunc

PInvokeMarshalToNativeFunc

ParameterIndex

PropertyIndex

RGCTXIndex

SequencePointKind

StringIndex

StringLiteralIndex

SynchronizationContextCallback

TypeDefinitionIndex

TypeEventIndex

TypeFieldIndex

TypeIndex

TypeInterfaceIndex

TypeInterfaceOffsetIndex

TypeMethodIndex

TypeNestedTypeIndex

TypePropertyIndex

VTableIndex

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__locale_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

il2cpp_array_lower_bound_t

il2cpp_array_size_t

il2cpp_hresult_t

il2cpp_liveness_reallocate_callback

il2cpp_register_object_callback

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

locale_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t

Unions

Il2CppDecimal__bindgen_ty_1

Il2CppDecimal__bindgen_ty_2

Il2CppDouble_double

Il2CppHStringHeader__bindgen_ty_1

Il2CppObject__bindgen_ty_1

Il2CppRGCTXData

Il2CppRGCTXDefinitionData

Il2CppSingle_float

Il2CppType__bindgen_ty_1

Il2CppVariant__bindgen_ty_1

Il2CppVariant__bindgen_ty_1___tagVARIANT__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_2

MethodInfo__bindgen_ty_1

MethodInfo__bindgen_ty_2