libil2cpp

Struct FieldInfo

pub struct FieldInfo(/* private fields */);

Information about a C# field

Implementations

impl FieldInfo

pub fn store<A>(&self, instance: &mut Il2CppObject, val: A)
where A: Argument,

Store a typechecked value into a field

pub unsafe fn store_unchecked<A>(&self, instance: &mut Il2CppObject, val: A)
where A: Argument,

Store a value into a field without type checking

Safety

To be safe, the provided type has to match the field signature

pub fn load<'a, T>(&self, instance: &'a mut Il2CppObject) -> T::Held<'a>
where T: Type,

Load a typechecked value from a field

pub unsafe fn load_unchecked<'a, T>( &self, instance: &'a mut Il2CppObject, ) -> T::Held<'a>
where T: Type,

Load a value from a field without type checking

Safety

To be safe, the provided type has to match the field signature

pub fn name(&self) -> Cow<'_, str>

Name of the field

pub fn parent(&self) -> &Il2CppClass

Class the field is from

pub fn ty(&self) -> &Il2CppType

Type of the field