libil2cpp

Struct Gc

pub struct Gc<T>(/* private fields */);

Wrapper type which implies the type is GC managed lifetime This is a Weak pointer to the GC managed object. If you want to hold a strong reference, use crate::GcBox<T> instead.

This is nullable, and can be used to represent a null reference to a GC managed object.

Implementations

impl<T> Gc<T>

pub fn new(ptr: *mut T) -> Self

Creates a new Gc instance with the given pointer.

pub fn null() -> Self

Creates a new Gc instance with a null pointer.

pub fn is_null(&self) -> bool

Checks if the pointer is null.

pub fn get_pointer(&self) -> *const T

Returns a constant pointer to the value.

pub fn get_pointer_mut(&mut self) -> *mut T

Returns a mutable pointer to the value.

pub fn as_ref(&self) -> Option<&T>

Returns an Option containing a reference to the value if the pointer is not null.

pub fn as_mut(&mut self) -> Option<&mut T>

Returns an Option containing a mutable reference to the value if the pointer is not null.

pub fn ensure_nonnull(&self) -> Result<NonNullGc<T>, NullGcError>

Ensures that the Gc<T> is not null, returning a NonNullGc<T> if it is not null, or a NullGcError if it is null.

pub fn type_cast<U>(self) -> Gc<U>
where T: AsMut<U>,

Casts T to U using compiler-checked type conversion, which will fail to compile if T is not convertible to U. This is a compile-time checked cast, and will not perform any runtime checks.

See cast for a runtime-checked cast that can fail at runtime.

impl<T> Gc<T>
where T: RefType,

pub fn cast<U>(self) -> Result<Gc<U>, String>
where U: RefType, T: RefType,

Casts T to U, checked against the object’s actual runtime class - can fail (Err) if it isn’t really (assignable to) a U. Compare type_cast, which trusts a compile-time relationship instead of checking and never fails.

If the Gc<T> is null, this will return a null Gc<U> regardless of the type relationship.

C++ Implementation https://github.com/QuestPackageManager/beatsaber-hook/blob/2604126ec26dd807da0be0ad974056d1f5fe9575/shared/utils/il2cpp-utils-classes.hpp#L185-L212

Safety

This function is safe to call, but the caller must ensure that the Gc<T> is valid and points to a valid object of type T. If the Gc<T> points to an invalid object, this function may cause undefined behavior.

pub fn into_safe_ptr(self) -> SafePtr<T>

Converts the current Gc instance to a SafePtr instance of the same type.