Struct libil2cpp::Il2CppObject

#[repr(transparent)]
pub struct Il2CppObject(_);

An il2cpp object

Implementations

impl Il2CppObject

pub fn class(&self) -> &'static Il2CppClass

Il2CppClass of the object

pub fn invoke<A, R, const N: usize>(
&mut self,
name: &str,
args: A
) -> Result<R, &mut Il2CppException> where
A: Arguments<N>,
R: Return,

Invokes the method with the given name on self using the given arguments, with type checking

Panics

This method will panic if a matching method can’t be found.

pub fn invoke_void<A, const N: usize>(
&mut self,
name: &str,
args: A
) -> Result<(), &mut Il2CppException> where
A: Arguments<N>,

Invokes the void method with the given name on self using the given arguments, with type checking

Panics

This method will panic if a matching method can’t be found.

pub fn load<R>(&mut self, field: &str) -> R where
R: Return,

Loads a value from a field of self with the given name, with type checking

Panics

This method will panic if the given field can’t be found

pub fn store<A>(&mut self, field: &str, value: A) where
A: Argument,

Stores a given value into a field of self with the given name, with type checking

Panics

This method will panic if the given field can’t be found