libil2cpp::thread

Function attached_invoke

pub unsafe fn attached_invoke<F, T>(f: F) -> T
where
    F: FnOnce() -> T,

Runs f with the current thread attached to il2cpp for the duration, detaching again once f returns (or panics) - see attach.

Safety

Same as attach: the calling thread must not already be attached.