libil2cpp

Trait Argument

pub unsafe trait Argument {
    type Type: Any;

    // Required methods
    fn matches(ty: &Il2CppType) -> bool;
    fn class() -> &'static Il2CppClass;
    fn invokable(&mut self) -> *mut c_void;
}

Trait implemented by types that can be used as C# method arguments

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Required Associated Types

type Type: Any

Normalized type of the argument, useful for caching

Required Methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# argument with the given Il2CppType to call a method

fn class() -> &'static Il2CppClass

Il2CppClass of the argument’s static type - used to rank candidate overloads by how close a match they are (see Il2CppClass::find_method’s use of param_distance) once more than one candidate matches

fn invokable(&mut self) -> *mut c_void

Returns an untyped pointer which can be used as a libil2cpp argument

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".