libil2cpp

Trait Arguments

pub unsafe trait Arguments<const N: usize> {
    type Type: Any;

    // Required methods
    fn matches(tys: &[&Il2CppType]) -> bool;
    fn classes() -> [&'static Il2CppClass; N];
    fn invokable(&mut self) -> [*mut c_void; N];

    // Provided method
    fn matches_method(method: &MethodInfo) -> bool { ... }
}

Trait implemented by types that can be used as a collection of C# method arguments

Note

You should most likely not be implementing this trait yourself

Safety

The implementation must be correct

Required Associated Types

type Type: Any

Normalized type of the arguments, useful for caching

Required Methods

fn matches(tys: &[&Il2CppType]) -> bool

Checks whether the type can be used as a C# argument collection against an explicit list of Il2CppTypes, rather than pulling them from a MethodInfo’s declared parameters directly like matches does - used by Il2CppClass::find_method to type-check a generic method’s parameters after substituting its own type parameters with concrete ones (mirrors beatsaber-hook’s ParameterMatch substituting into a paramType before calling IsConvertibleFrom, il2cpp-utils-methods.hpp).

fn classes() -> [&'static Il2CppClass; N]

Il2CppClasses of each argument’s static type, in order - see Argument::class

fn invokable(&mut self) -> [*mut c_void; N]

Returns an array of untyped pointer which can be used to invoke C# methods

Provided Methods

fn matches_method(method: &MethodInfo) -> bool

Checks whether the type can be used as a C# argument collection for the given MethodInfo

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".