Trait libil2cpp::CalleeArgument

pub unsafe trait CalleeArgument {
    type Type: Any;
    fn matches(ty: &Il2CppType) -> bool;
}

Trait implemented by types that can be used as C# method parameters

Safety

Interfaces depending on this trait assume that all of its methods are correct in an il2cpp context

Associated Types

type Type: Any

Normalized type of the parameter, useful for caching

Required methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# parameter with the given Il2CppType in a method definition