Trait libil2cpp::CalleeReturn

pub unsafe trait CalleeReturn {
    type Type: Any;
    fn matches(ty: &Il2CppType) -> bool;
}

Trait implemented by types that can be used as callee return types for C# methods

Safety

Interfaces depending on this trait assume that all of its methods are correct in an il2cpp context

Associated Types

type Type: Any

Normalized type of the return type, useful for caching

Required methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# return type with the given Il2CppType in a method definition