Trait libil2cpp::ObjectExt

pub trait ObjectExt: for<'a> Type<Held<'a> = Option<&'a mut Self>> + DerefMut<Target = Il2CppObject> + Sized {
    fn new<A, const N: usize>(args: A) -> &'static mut Self
    where
        A: Arguments<N>,
    { ... }
}

Helper trait for reference types which can be dereferenced to an object

Provided Methods

fn new<A, const N: usize>(args: A) -> &'static mut Self where
A: Arguments<N>,

Creates a new object using the constructor taking the given arguments