libil2cpp

Trait ObjectType

pub trait ObjectType {
    // Required methods
    fn as_object(&self) -> &Il2CppObject;
    fn as_object_mut(&mut self) -> &mut Il2CppObject;
}

Trait for types that can be treated as Il2CppObject Useful for generic constraints and supporting multiple object-like types

Required Methods

fn as_object(&self) -> &Il2CppObject

Returns a reference to the underlying Il2CppObject

fn as_object_mut(&mut self) -> &mut Il2CppObject

Returns a mutable reference to the underlying Il2CppObject

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".