libil2cpp

Trait Parameter

pub unsafe trait Parameter {
    type Actual;

    // Required methods
    fn matches(ty: &Il2CppType) -> bool;
    fn class() -> &'static Il2CppClass;
    fn from_actual(actual: Self::Actual) -> Self;
    fn into_actual(self) -> Self::Actual;
}

Trait implemented by types that can be used as C# method parameters

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Required Associated Types

type Actual

Type of the actual parameter

Required Methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# parameter with the given Il2CppType

fn class() -> &'static Il2CppClass

Il2CppClass of the parameter’s static type - used to rank candidate overloads by how close a match they are, mirroring Argument::class

fn from_actual(actual: Self::Actual) -> Self

Converts from the actual type to the desired one

fn into_actual(self) -> Self::Actual

Converts from the desired type into the actual one

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".