libil2cpp

Trait Parameters

pub unsafe trait Parameters {
    const COUNT: usize;

    // Required method
    fn matches(method: &MethodInfo) -> bool;
}

Trait implemented by types that can be used as a collection of C# method parameters

Note

You should most likely not be implementing this trait yourself

Safety

The implementation must be correct

Required Associated Constants

const COUNT: usize

Parameter count

Required Methods

fn matches(method: &MethodInfo) -> bool

Checks whether the type can be used as a C# parameter collection for the given MethodInfo

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".