libil2cpp

Trait Return

pub unsafe trait Return {
    type Actual;

    // Required methods
    fn matches(ty: &Il2CppType) -> bool;
    fn into_actual(self) -> Self::Actual;
    fn from_actual(actual: Self::Actual) -> Self;
}

Trait implemented by types that can be used as return types for C# methods

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Required Associated Types

type Actual

Type of the actual return value

Required Methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# return type of the given Il2CppType

fn into_actual(self) -> Self::Actual

Converts from the desired type into the actual one

fn from_actual(actual: Self::Actual) -> Self

Converts from the actual type to the desired one

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".