Crate libil2cpp

Wrappers and raw bindings for Unity’s libil2cpp

Re-exports

pub use raw::unbox;

pub use raw::value_box_alloc;

pub use raw::WrapRaw;

pub use raw::IL2CPP_BINARY;

Modules

debug

Bulk introspection helpers for exploring the loaded il2cpp domain - ports beatsaber-hook’s debug.hpp (log_classes/log_methods/log_fields/ log_properties), useful for reverse-engineering during mod development. Per-item Debug/Display impls already exist on Il2CppClass/[MethodInfo]/[FieldInfo]/ [PropertyInfo]; what’s missing without this module is a way to walk all of them at once.

raw

Raw il2cpp types and functions

thread

Helpers for running Rust code on native threads that talk to il2cpp.

Macros

unsafe_impl_reference_type

Implements the Type trait for a Rust type that is equivalent to a C# reference type

unsafe_impl_value_type

Implements the Type trait for a Rust type that is equivalent to a C# value type

Structs

BoxedValue

A boxed C# value type sitting on the GC heap - the result of boxing a ValueType (see crate::raw::value_box_alloc).

ByRef

TODO: Add lifetime param Wrapper type which implies the type is ByRef managed lifetime

FieldInfo

Information about a C# field

Gc

Wrapper type which implies the type is GC managed lifetime This is a Weak pointer to the GC managed object. If you want to hold a strong reference, use crate::GcBox<T> instead.

GcAllocator

An allocator that uses GC functions to allocate memory. This is useful for allocating memory that will be managed by the GC, such as objects that will be used in the Unity engine.

Il2CppArray

An il2cpp array

Il2CppClass

An il2cpp class

Il2CppException

An il2cpp exception

Il2CppObject

An il2cpp object

Il2CppReflectionMethod

Object used for reflection of methods

Il2CppReflectionType

Object used for reflection of types

Il2CppString

An il2cpp string

Il2CppType

An il2cpp type

MethodInfo

Information about a C# method

NonNullGc

A Gc<T> that is statically known to never be null.

ParameterInfo

Information about a C# parameter

PropertyInfo

Information about a C# property

SafePtr

A strong reference to a GC-managed il2cpp object, safe to hold onto across GC collections (unlike Gc<T>, a weak, untracked pointer whose pointee the GC may collect at any time) and safe to share across threads.

ValueTypePadding

Padding type for value types

Enums

Builtin

Builtin C# types

FindMethodError

Possible errors when looking up a method

Traits

Argument

Trait implemented by types that can be used as C# method arguments

Arguments

Trait implemented by types that can be used as a collection of C# method arguments

Generics

Trait implemented for Rust types which can represent a list of C# generic arguments

ObjectExt

Helper trait for reference types which can be dereferenced to an object

Parameter

Trait implemented by types that can be used as C# method parameters

Parameters

Trait implemented by types that can be used as a collection of C# method parameters

RefType

Marker trait for C# reference types - types held via a nullable pointer (Held<'a> = Option<&'a mut Self>), as opposed to crate::ValueTypes, which are held by value (Held<'a> = Self). A single concrete type can only satisfy one of the two, since Type::Held can’t be both at once.

Return

Trait implemented by types that can be used as return types for C# methods

Returned

Trait implemented by types that can be used as return types from C# methods

ThisArgument

Trait implemented by types that can be used as a C# this arguments

ThisParameter

Trait implemented by types that can be used as C# this method parameters

Type

Trait implemented by Rust types that are also C# types

ValueType

Extension trait for value types providing additional functionality

Type Aliases

ByRefMut

Mutable ByRef type alias

GcArc

Arc with GcAllocator.

GcBox

Box with GcAllocator.

GcHashMap

HashMap with GcAllocator.

GcRc

Rc with GcAllocator.

GcVec

Vec with GcAllocator.

Result

Result type for method invocations

Void

Alias for () to represent void return type