Crate libil2cpp

Wrappers and raw bindings for Unity’s libil2cpp

Re-exports

pub use raw::unbox;

pub use raw::WrapRaw;

Modules

raw

Raw il2cpp types and functions

Macros

unsafe_impl_reference_type

Implements the Type trait for a C# reference type

unsafe_impl_value_type

Implements the Type trait for a C# value type

Structs

FieldInfo

Information about a C# field

Il2CppArray

An il2cpp array

Il2CppClass

An il2cpp class

Il2CppException

An il2cpp exception

Il2CppObject

An il2cpp object

Il2CppReflectionMethod

Object used for reflection of methods

Il2CppReflectionType

Object used for reflection of types

Il2CppString

An il2cpp string

Il2CppType

An il2cpp type

MethodInfo

Information about a C# method

ParameterInfo

Information about a C# parameter

Reference

Marker type used to specify reference semantics

Value

Marker type used to specify value semantics

Enums

Builtin

Builtin C# types

FindMethodError

Possible errors when looking up a method

Traits

Argument

Trait implemented by types that can be used as C# method arguments

Arguments

Trait implemented by types that can be used as a collection of C# method arguments

Parameter

Trait implemented by types that can be used as C# method parameters

Parameters

Trait implemented by types that can be used as a collection of C# method parameters

Return

Trait implemented by types that can be used as return types for C# methods

Returned

Trait implemented by types that can be used as return types from C# methods

ThisArgument

Trait implemented by types that can be used as a C# this arguments

ThisParameter

Trait implemented by types that can be used as C# this method parameters

Type

Trait implemented by Rust types that are also C# types