Crate libil2cpp

Re-exports

pub use raw::WrapRaw;

Modules

raw

Raw il2cpp types and functions

Structs

FieldInfo

Information about a C# field

Il2CppArray

An il2cpp array

Il2CppClass

An il2cpp class

Il2CppException

An il2cpp object

Il2CppObject

An il2cpp object

Il2CppString

An il2cpp string

Il2CppType

An il2cpp type

MethodInfo

Information about a C# method

ParameterInfo

Information about a C# parameter

Enums

Builtin

Traits

Argument

Trait implemented by types that can be used as C# method arguments

Arguments

Trait implemented by types that can be used as a collection of C# method arguments

CalleeArgument

Trait implemented by types that can be used as C# method parameters

CalleeArguments

Trait implemented by types that can be used as a collection of C# method parameters

CalleeReturn

Trait implemented by types that can be used as callee return types for C# methods

CalleeThis

Trait implemented by types that can be used as a C# instance parameters

Parameter

Trait implemented by types that can be used as C# method parameters

Parameters

Trait implemented by types that can be used as a collection of C# method parameters

Return

Trait implemented by types that can be used as caller return types for C# methods

This

Trait implemented by types that can be used as a C# instance arguments

Type

Trait implemented by Rust types that are also C# types