Module libil2cpp::raw

Raw il2cpp types and functions

This module contains raw C types defined in libil2cpp and raw C functions dynamically loaded from libil2cpp.so.

Structs

CustomAttributesCache

EventInfo

FieldInfo

Il2CppAppContext

Il2CppAppDomain

Il2CppAppDomainSetup

Il2CppArgumentException

Il2CppArray

Il2CppArrayBounds

Il2CppArraySize

Il2CppArrayType

Il2CppAssembly

Il2CppAssemblyDefinition

Il2CppAssemblyName

Il2CppAssemblyNameDefinition

Il2CppAsyncCall

Il2CppAsyncResult

Il2CppCalendarData

Il2CppCatchPoint

Il2CppClass

Il2CppCodeGenModule

Il2CppCodeGenOptions

Il2CppCodeRegistration

Il2CppCultureData

Il2CppCultureInfo

Il2CppCustomAttrInfo

Il2CppCustomAttributeTypeRange

Il2CppDateTimeFormatInfo

Il2CppDebuggerMetadataRegistration

Il2CppDebuggerTransport

Il2CppDecimal

Il2CppDecimal__bindgen_ty_1__bindgen_ty_1

Il2CppDecimal__bindgen_ty_2__bindgen_ty_1

Il2CppDefaults

Il2CppDelegate

Il2CppDomain

Il2CppDouble

Il2CppErrorWrapper

Il2CppEventDefinition

Il2CppException

Il2CppExceptionWrapper

Il2CppFieldDefaultValue

Il2CppFieldDefinition

Il2CppFieldMarshaledSize

Il2CppFieldRef

Il2CppFileTime

Il2CppGenericClass

Il2CppGenericContainer

Il2CppGenericContext

Il2CppGenericInst

Il2CppGenericMethod

Il2CppGenericMethodFunctionsDefinitions

Il2CppGenericMethodIndices

Il2CppGenericParameter

Il2CppGlobalMetadataHeader

Il2CppGuid

Il2CppHStringHeader

Il2CppHString__

Il2CppIOAsyncResult

Il2CppIOSelectorJob

Il2CppIUnknown

Il2CppImage

Il2CppImageDefinition

Il2CppInterfaceOffsetPair

Il2CppInternalThread

Il2CppInteropData

Il2CppManagedMemorySnapshot

Il2CppManifestResourceInfo

Il2CppMarshalByRefObject

Il2CppMemoryCallbacks

Il2CppMetadataRegistration

Il2CppMetadataUsageList

Il2CppMetadataUsagePair

Il2CppMethodDebugInfo

Il2CppMethodDefinition

Il2CppMethodExecutionContextInfo

Il2CppMethodExecutionContextInfoIndex

Il2CppMethodHeaderInfo

Il2CppMethodInfo

Il2CppMethodMessage

Il2CppMethodScope

Il2CppMethodSpec

Il2CppMulticastDelegate

Il2CppNameToTypeDefinitionIndexHashTable

Il2CppNumberFormatInfo

Il2CppObject

Il2CppParameterDefaultValue

Il2CppParameterDefinition

Il2CppPerfCounters

Il2CppProfiler

Il2CppPropertyDefinition

Il2CppPropertyInfo

Il2CppRGCTXDefinition

Il2CppRange

Il2CppReflectionAssembly

Il2CppReflectionAssemblyName

Il2CppReflectionEvent

Il2CppReflectionField

Il2CppReflectionGenericMethod

Il2CppReflectionMarshal

Il2CppReflectionMethod

Il2CppReflectionModule

Il2CppReflectionMonoEvent

Il2CppReflectionMonoEventInfo

Il2CppReflectionMonoType

Il2CppReflectionParameter

Il2CppReflectionPointer

Il2CppReflectionProperty

Il2CppReflectionRuntimeType

Il2CppReflectionType

Il2CppRegionInfo

Il2CppRuntimeInterfaceOffsetPair

Il2CppSafeArray

Il2CppSafeArrayBound

Il2CppSafeHandle

Il2CppSequencePoint

Il2CppSequencePointSourceFile

Il2CppSingle

Il2CppSocketAddress

Il2CppSocketAsyncResult

Corresponds to Mono’s internal System.Net.Sockets.Socket.SocketAsyncResult class. Has no relation to Il2CppAsyncResult.

Il2CppSortKey

Il2CppStackFrame

Il2CppStackFrameInfo

Il2CppStatStg

Il2CppString

Il2CppStringBuilder

Il2CppStringLiteral

Il2CppSystemException

Il2CppThread

Il2CppTokenIndexMethodTuple

Il2CppTokenIndexPair

Il2CppTokenRangePair

Il2CppType

Il2CppTypeDefinition

Il2CppTypeDefinitionMetadata

Il2CppTypeDefinitionSizes

Il2CppTypeSourceFilePair

Il2CppTypedRef

Il2CppVariant

Il2CppVariant__bindgen_ty_1___tagVARIANT

Il2CppVariant__bindgen_ty_1___tagVARIANT__bindgen_ty_1___tagBRECORD

Il2CppWaitHandle

Il2CppWin32Decimal

Il2CppWin32Decimal__bindgen_ty_1__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_2__bindgen_ty_1

Il2CppWindowsRuntimeFactoryTableEntry

Il2CppWindowsRuntimeTypeName

Il2CppWindowsRuntimeTypeNamePair

MemberInfo

MethodInfo

MonitorData

ParameterInfo

PropertyInfo

VirtualInvokeData

_G_fpos64_t

_G_fpos_t

_IO_FILE

_IO_codecvt

_IO_marker

_IO_wide_data

__BindgenBitfieldUnit

__IncompleteArrayField

__fsid_t

__locale_data

__locale_struct

__mbstate_t

__once_flag

__pthread_cond_s

__pthread_cond_s__bindgen_ty_1__bindgen_ty_1

__pthread_cond_s__bindgen_ty_2__bindgen_ty_1

__pthread_internal_list

__pthread_internal_slist

__pthread_mutex_s

__pthread_rwlock_arch_t

__sigset_t

__va_list_tag

div_t

drand48_data

fd_set

ldiv_t

lldiv_t

max_align_t

random_data

timespec

timeval

Constants

ASSEMBLYREF_DISABLEJITCOMPILE_OPTIMIZER_FLAG

ASSEMBLYREF_ENABLEJITCOMPILE_TRACKING_FLAG

ASSEMBLYREF_FULL_PUBLIC_KEY_FLAG

ASSEMBLYREF_RETARGETABLE_FLAG

BIG_ENDIAN

BUFSIZ

BYTE_ORDER

EOF

EXIT_FAILURE

EXIT_SUCCESS

FD_SETSIZE

FIELD_ATTRIBUTE_ASSEMBLY

FIELD_ATTRIBUTE_COMPILER_CONTROLLED

FIELD_ATTRIBUTE_FAMILY

FIELD_ATTRIBUTE_FAM_AND_ASSEM

FIELD_ATTRIBUTE_FAM_OR_ASSEM

FIELD_ATTRIBUTE_FIELD_ACCESS_MASK

FIELD_ATTRIBUTE_HAS_DEFAULT

FIELD_ATTRIBUTE_HAS_FIELD_MARSHAL

FIELD_ATTRIBUTE_HAS_FIELD_RVA

FIELD_ATTRIBUTE_INIT_ONLY

FIELD_ATTRIBUTE_LITERAL

FIELD_ATTRIBUTE_NOT_SERIALIZED

FIELD_ATTRIBUTE_PINVOKE_IMPL

FIELD_ATTRIBUTE_PRIVATE

FIELD_ATTRIBUTE_PUBLIC

FIELD_ATTRIBUTE_RESERVED_MASK

FIELD_ATTRIBUTE_RT_SPECIAL_NAME

FIELD_ATTRIBUTE_SPECIAL_NAME

FIELD_ATTRIBUTE_STATIC

FILENAME_MAX

FIXME

FOPEN_MAX

IL2CPP_API_DYNAMIC_NO_DLSYM

IL2CPP_BIG_ENDIAN

IL2CPP_BINARY_SECTION_NAME

IL2CPP_BYTE_ORDER

IL2CPP_CAN_USE_MULTIPLE_SYMBOL_MAPS

IL2CPP_CXX_ABI_MSVC

IL2CPP_DEBUG

IL2CPP_DEVELOPMENT

IL2CPP_DIR_SEPARATOR

IL2CPP_DISABLE_FULL_MESSAGES

IL2CPP_ENABLE_DEFERRED_GC

IL2CPP_ENABLE_MONO_BUG_EMULATION

IL2CPP_ENABLE_PLATFORM_THREAD_RENAME

IL2CPP_ENABLE_STACKTRACES

IL2CPP_EXCEPTION_DISABLED

IL2CPP_GC_BOEHM

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_CONTRAVARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_COVARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_DEFAULT_CONSTRUCTOR_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_NON_VARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_NOT_NULLABLE_VALUE_TYPE_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_REFERENCE_TYPE_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_SPECIAL_CONSTRAINT_MASK

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_VARIANCE_MASK

IL2CPP_HAS_DELETED_FUNCTIONS

IL2CPP_HAS_GC_DESCRIPTORS

IL2CPP_LARGE_EXECUTABLE_ARM_WORKAROUND

IL2CPP_LITTLE_ENDIAN

IL2CPP_PAGE_SIZE

IL2CPP_PLATFORM_DISABLE_LIBC_PINVOKE

IL2CPP_PLATFORM_OVERRIDES_STD_FILE_HANDLES

IL2CPP_PLATFORM_SUPPORTS_CPU_INFO

IL2CPP_PLATFORM_SUPPORTS_DEBUGGER_PRESENT

IL2CPP_PLATFORM_SUPPORTS_SYSTEM_CERTIFICATES

IL2CPP_PLATFORM_SUPPORTS_TIMEZONEINFO

IL2CPP_POINTER_SPARE_BITS

IL2CPP_SIZEOF_STRUCT_WITH_NO_INSTANCE_FIELDS

IL2CPP_SIZEOF_VOID_P

IL2CPP_TARGET_ANDROID

IL2CPP_TARGET_ARM64

IL2CPP_TARGET_ARMV7

IL2CPP_TARGET_DARWIN

IL2CPP_TARGET_IOS

IL2CPP_TARGET_JAVASCRIPT

IL2CPP_TARGET_LINUX

IL2CPP_TARGET_LUMIN

IL2CPP_TARGET_N3DS

IL2CPP_TARGET_OSX

IL2CPP_TARGET_PS4

IL2CPP_TARGET_PSP2

IL2CPP_TARGET_SWITCH

IL2CPP_TARGET_WINDOWS

IL2CPP_TARGET_WINDOWS_DESKTOP

IL2CPP_TARGET_WINDOWS_GAMES

IL2CPP_TARGET_WINRT

IL2CPP_TARGET_XBOXONE

IL2CPP_THREADS_STD

IL2CPP_THREAD_IMPL_HAS_COM_APARTMENTS

IL2CPP_USE_GENERIC_CPU_INFO

IL2CPP_USE_SEND_NOSIGNAL

IL2CPP_USE_STD_THREAD

IL2CPP_VALIDATE_FIELD_LAYOUT

IL2CPP_ZERO_LEN_ARRAY

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

Il2CppCallConvention_IL2CPP_CALL_C

Il2CppCallConvention_IL2CPP_CALL_DEFAULT

Il2CppCallConvention_IL2CPP_CALL_FASTCALL

Il2CppCallConvention_IL2CPP_CALL_STDCALL

Il2CppCallConvention_IL2CPP_CALL_THISCALL

Il2CppCallConvention_IL2CPP_CALL_VARARG

Il2CppCallType_Il2Cpp_CallType_BeginInvoke

Il2CppCallType_Il2Cpp_CallType_EndInvoke

Il2CppCallType_Il2Cpp_CallType_OneWay

Il2CppCallType_Il2Cpp_CallType_Sync

Il2CppCharSet_CHARSET_ANSI

Il2CppCharSet_CHARSET_NOT_SPECIFIED

Il2CppCharSet_CHARSET_UNICODE

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_EQ

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_GT

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_LT

Il2CppGCEvent_IL2CPP_GC_EVENT_END

Il2CppGCEvent_IL2CPP_GC_EVENT_MARK_END

Il2CppGCEvent_IL2CPP_GC_EVENT_MARK_START

Il2CppGCEvent_IL2CPP_GC_EVENT_POST_START_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_POST_STOP_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_PRE_START_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_PRE_STOP_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_RECLAIM_END

Il2CppGCEvent_IL2CPP_GC_EVENT_RECLAIM_START

Il2CppGCEvent_IL2CPP_GC_EVENT_START

Il2CppMetadataUsage_kIl2CppMetadataUsageFieldInfo

Il2CppMetadataUsage_kIl2CppMetadataUsageIl2CppType

Il2CppMetadataUsage_kIl2CppMetadataUsageInvalid

Il2CppMetadataUsage_kIl2CppMetadataUsageMethodDef

Il2CppMetadataUsage_kIl2CppMetadataUsageMethodRef

Il2CppMetadataUsage_kIl2CppMetadataUsageStringLiteral

Il2CppMetadataUsage_kIl2CppMetadataUsageTypeInfo

Il2CppProfileFileIOKind_IL2CPP_PROFILE_FILEIO_READ

Il2CppProfileFileIOKind_IL2CPP_PROFILE_FILEIO_WRITE

Il2CppProfileFlags_IL2CPP_PROFILE_ALLOCATIONS

Il2CppProfileFlags_IL2CPP_PROFILE_APPDOMAIN_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_ASSEMBLY_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_CLASS_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_COVERAGE

Il2CppProfileFlags_IL2CPP_PROFILE_ENTER_LEAVE

Il2CppProfileFlags_IL2CPP_PROFILE_EXCEPTIONS

Il2CppProfileFlags_IL2CPP_PROFILE_FILEIO

Il2CppProfileFlags_IL2CPP_PROFILE_GC

Il2CppProfileFlags_IL2CPP_PROFILE_GC_MOVES

Il2CppProfileFlags_IL2CPP_PROFILE_INLINING

Il2CppProfileFlags_IL2CPP_PROFILE_INS_COVERAGE

Il2CppProfileFlags_IL2CPP_PROFILE_IOMAP_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_JIT_COMPILATION

Il2CppProfileFlags_IL2CPP_PROFILE_METHOD_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_MODULE_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_MONITOR_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_NONE

Il2CppProfileFlags_IL2CPP_PROFILE_REMOTING

Il2CppProfileFlags_IL2CPP_PROFILE_STATISTICAL

Il2CppProfileFlags_IL2CPP_PROFILE_THREADS

Il2CppProfileFlags_IL2CPP_PROFILE_TRANSITIONS

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_ARRAY

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_CLASS

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_INVALID

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_METHOD

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_TYPE

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_ANOTHER_ASSEMBLY

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_EMBEDDED

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_IN_MANIFEST

Il2CppRuntimeUnhandledExceptionPolicy_IL2CPP_UNHANDLED_POLICY_CURRENT

Il2CppRuntimeUnhandledExceptionPolicy_IL2CPP_UNHANDLED_POLICY_LEGACY

Il2CppStat_IL2CPP_STAT_CLASS_STATIC_DATA_SIZE

Il2CppStat_IL2CPP_STAT_GENERIC_CLASS_COUNT

Il2CppStat_IL2CPP_STAT_GENERIC_INSTANCE_COUNT

Il2CppStat_IL2CPP_STAT_INFLATED_METHOD_COUNT

Il2CppStat_IL2CPP_STAT_INFLATED_TYPE_COUNT

Il2CppStat_IL2CPP_STAT_INITIALIZED_CLASS_COUNT

Il2CppStat_IL2CPP_STAT_METHOD_COUNT

Il2CppStat_IL2CPP_STAT_NEW_OBJECT_COUNT

Il2CppTokenType_IL2CPP_TOKEN_ASSEMBLY

Il2CppTokenType_IL2CPP_TOKEN_ASSEMBLY_REF

Il2CppTokenType_IL2CPP_TOKEN_CUSTOM_ATTRIBUTE

Il2CppTokenType_IL2CPP_TOKEN_EVENT

Il2CppTokenType_IL2CPP_TOKEN_EXPORTED_TYPE

Il2CppTokenType_IL2CPP_TOKEN_FIELD_DEF

Il2CppTokenType_IL2CPP_TOKEN_FILE

Il2CppTokenType_IL2CPP_TOKEN_GENERIC_PARAM

Il2CppTokenType_IL2CPP_TOKEN_INTERFACE_IMPL

Il2CppTokenType_IL2CPP_TOKEN_MANIFEST_RESOURCE

Il2CppTokenType_IL2CPP_TOKEN_MEMBER_REF

Il2CppTokenType_IL2CPP_TOKEN_METHOD_DEF

Il2CppTokenType_IL2CPP_TOKEN_METHOD_SPEC

Il2CppTokenType_IL2CPP_TOKEN_MODULE

Il2CppTokenType_IL2CPP_TOKEN_MODULE_REF

Il2CppTokenType_IL2CPP_TOKEN_PARAM_DEF

Il2CppTokenType_IL2CPP_TOKEN_PERMISSION

Il2CppTokenType_IL2CPP_TOKEN_PROPERTY

Il2CppTokenType_IL2CPP_TOKEN_SIGNATURE

Il2CppTokenType_IL2CPP_TOKEN_TYPE_DEF

Il2CppTokenType_IL2CPP_TOKEN_TYPE_REF

Il2CppTokenType_IL2CPP_TOKEN_TYPE_SPEC

Il2CppTypeEnum_IL2CPP_TYPE_ARRAY

Il2CppTypeEnum_IL2CPP_TYPE_BOOLEAN

Il2CppTypeEnum_IL2CPP_TYPE_BYREF

Il2CppTypeEnum_IL2CPP_TYPE_CHAR

Il2CppTypeEnum_IL2CPP_TYPE_CLASS

Il2CppTypeEnum_IL2CPP_TYPE_CMOD_OPT

Il2CppTypeEnum_IL2CPP_TYPE_CMOD_REQD

Il2CppTypeEnum_IL2CPP_TYPE_END

Il2CppTypeEnum_IL2CPP_TYPE_ENUM

Il2CppTypeEnum_IL2CPP_TYPE_FNPTR

Il2CppTypeEnum_IL2CPP_TYPE_GENERICINST

Il2CppTypeEnum_IL2CPP_TYPE_I

Il2CppTypeEnum_IL2CPP_TYPE_I1

Il2CppTypeEnum_IL2CPP_TYPE_I2

Il2CppTypeEnum_IL2CPP_TYPE_I4

Il2CppTypeEnum_IL2CPP_TYPE_I8

Il2CppTypeEnum_IL2CPP_TYPE_INTERNAL

Il2CppTypeEnum_IL2CPP_TYPE_MODIFIER

Il2CppTypeEnum_IL2CPP_TYPE_MVAR

Il2CppTypeEnum_IL2CPP_TYPE_OBJECT

Il2CppTypeEnum_IL2CPP_TYPE_PINNED

Il2CppTypeEnum_IL2CPP_TYPE_PTR

Il2CppTypeEnum_IL2CPP_TYPE_R4

Il2CppTypeEnum_IL2CPP_TYPE_R8

Il2CppTypeEnum_IL2CPP_TYPE_SENTINEL

Il2CppTypeEnum_IL2CPP_TYPE_STRING

Il2CppTypeEnum_IL2CPP_TYPE_SZARRAY

Il2CppTypeEnum_IL2CPP_TYPE_TYPEDBYREF

Il2CppTypeEnum_IL2CPP_TYPE_U

Il2CppTypeEnum_IL2CPP_TYPE_U1

Il2CppTypeEnum_IL2CPP_TYPE_U2

Il2CppTypeEnum_IL2CPP_TYPE_U4

Il2CppTypeEnum_IL2CPP_TYPE_U8

Il2CppTypeEnum_IL2CPP_TYPE_VALUETYPE

Il2CppTypeEnum_IL2CPP_TYPE_VAR

Il2CppTypeEnum_IL2CPP_TYPE_VOID

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_ASSEMBLY_QUALIFIED

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_FULL_NAME

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_IL

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_REFLECTION

Il2CppVarType_IL2CPP_VT_ARRAY

Il2CppVarType_IL2CPP_VT_BLOB

Il2CppVarType_IL2CPP_VT_BLOB_OBJECT

Il2CppVarType_IL2CPP_VT_BOOL

Il2CppVarType_IL2CPP_VT_BSTR

Il2CppVarType_IL2CPP_VT_BSTR_BLOB

Il2CppVarType_IL2CPP_VT_BYREF

Il2CppVarType_IL2CPP_VT_CARRAY

Il2CppVarType_IL2CPP_VT_CF

Il2CppVarType_IL2CPP_VT_CLSID

Il2CppVarType_IL2CPP_VT_CY

Il2CppVarType_IL2CPP_VT_DATE

Il2CppVarType_IL2CPP_VT_DECIMAL

Il2CppVarType_IL2CPP_VT_DISPATCH

Il2CppVarType_IL2CPP_VT_EMPTY

Il2CppVarType_IL2CPP_VT_ERROR

Il2CppVarType_IL2CPP_VT_FILETIME

Il2CppVarType_IL2CPP_VT_HRESULT

Il2CppVarType_IL2CPP_VT_I1

Il2CppVarType_IL2CPP_VT_I2

Il2CppVarType_IL2CPP_VT_I4

Il2CppVarType_IL2CPP_VT_I8

Il2CppVarType_IL2CPP_VT_ILLEGAL

Il2CppVarType_IL2CPP_VT_ILLEGALMASKED

Il2CppVarType_IL2CPP_VT_INT

Il2CppVarType_IL2CPP_VT_INT_PTR

Il2CppVarType_IL2CPP_VT_LPSTR

Il2CppVarType_IL2CPP_VT_LPWSTR

Il2CppVarType_IL2CPP_VT_NULL

Il2CppVarType_IL2CPP_VT_PTR

Il2CppVarType_IL2CPP_VT_R4

Il2CppVarType_IL2CPP_VT_R8

Il2CppVarType_IL2CPP_VT_RECORD

Il2CppVarType_IL2CPP_VT_RESERVED

Il2CppVarType_IL2CPP_VT_SAFEARRAY

Il2CppVarType_IL2CPP_VT_STORAGE

Il2CppVarType_IL2CPP_VT_STORED_OBJECT

Il2CppVarType_IL2CPP_VT_STREAM

Il2CppVarType_IL2CPP_VT_STREAMED_OBJECT

Il2CppVarType_IL2CPP_VT_TYPEMASK

Il2CppVarType_IL2CPP_VT_UI1

Il2CppVarType_IL2CPP_VT_UI2

Il2CppVarType_IL2CPP_VT_UI4

Il2CppVarType_IL2CPP_VT_UI8

Il2CppVarType_IL2CPP_VT_UINT

Il2CppVarType_IL2CPP_VT_UINT_PTR

Il2CppVarType_IL2CPP_VT_UNKNOWN

Il2CppVarType_IL2CPP_VT_USERDEFINED

Il2CppVarType_IL2CPP_VT_VARIANT

Il2CppVarType_IL2CPP_VT_VECTOR

Il2CppVarType_IL2CPP_VT_VERSIONED_STREAM

Il2CppVarType_IL2CPP_VT_VOID

Il2CppWindowsRuntimeTypeKind_kTypeKindCustom

Il2CppWindowsRuntimeTypeKind_kTypeKindMetadata

Il2CppWindowsRuntimeTypeKind_kTypeKindPrimitive

LITTLE_ENDIAN

L_ctermid

L_tmpnam

MAXIMUM_NESTED_GENERICS_EXCEPTION_MESSAGE

METHOD_ATTRIBUTE_ABSTRACT

METHOD_ATTRIBUTE_ASSEM

METHOD_ATTRIBUTE_COMPILER_CONTROLLED

METHOD_ATTRIBUTE_FAMILY

METHOD_ATTRIBUTE_FAM_AND_ASSEM

METHOD_ATTRIBUTE_FAM_OR_ASSEM

METHOD_ATTRIBUTE_FINAL

METHOD_ATTRIBUTE_HAS_SECURITY

METHOD_ATTRIBUTE_HIDE_BY_SIG

METHOD_ATTRIBUTE_MEMBER_ACCESS_MASK

METHOD_ATTRIBUTE_NEW_SLOT

METHOD_ATTRIBUTE_PINVOKE_IMPL

METHOD_ATTRIBUTE_PRIVATE

METHOD_ATTRIBUTE_PUBLIC

METHOD_ATTRIBUTE_REQUIRE_SEC_OBJECT

METHOD_ATTRIBUTE_RESERVED_MASK

METHOD_ATTRIBUTE_REUSE_SLOT

METHOD_ATTRIBUTE_RT_SPECIAL_NAME

METHOD_ATTRIBUTE_SPECIAL_NAME

METHOD_ATTRIBUTE_STATIC

METHOD_ATTRIBUTE_STRICT

METHOD_ATTRIBUTE_UNMANAGED_EXPORT

METHOD_ATTRIBUTE_VIRTUAL

METHOD_ATTRIBUTE_VTABLE_LAYOUT_MASK

METHOD_IMPL_ATTRIBUTE_CODE_TYPE_MASK

METHOD_IMPL_ATTRIBUTE_FORWARD_REF

METHOD_IMPL_ATTRIBUTE_IL

METHOD_IMPL_ATTRIBUTE_INTERNAL_CALL

METHOD_IMPL_ATTRIBUTE_MANAGED

METHOD_IMPL_ATTRIBUTE_MANAGED_MASK

METHOD_IMPL_ATTRIBUTE_MAX_METHOD_IMPL_VAL

METHOD_IMPL_ATTRIBUTE_NATIVE

METHOD_IMPL_ATTRIBUTE_NOINLINING

METHOD_IMPL_ATTRIBUTE_OPTIL

METHOD_IMPL_ATTRIBUTE_PRESERVE_SIG

METHOD_IMPL_ATTRIBUTE_RUNTIME

METHOD_IMPL_ATTRIBUTE_SYNCHRONIZED

METHOD_IMPL_ATTRIBUTE_UNMANAGED

MethodVariableKind_kMethodVariableKind_LocalVariable

MethodVariableKind_kMethodVariableKind_Parameter

MethodVariableKind_kMethodVariableKind_This

NEED_TO_ZERO_PTRFREE

NET_4_0

PARAM_ATTRIBUTE_HAS_DEFAULT

PARAM_ATTRIBUTE_HAS_FIELD_MARSHAL

PARAM_ATTRIBUTE_IN

PARAM_ATTRIBUTE_OPTIONAL

PARAM_ATTRIBUTE_OUT

PARAM_ATTRIBUTE_RESERVED_MASK

PARAM_ATTRIBUTE_UNUSED

PDP_ENDIAN

PTRDIFF_MAX

PTRDIFF_MIN

PUBLIC_KEY_BYTE_LENGTH

P_tmpdir

RAND_MAX

SEEK_CUR

SEEK_END

SEEK_SET

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

SequencePointKind_kSequencePointKind_Normal

SequencePointKind_kSequencePointKind_StepOut

THREAD_STATIC_FIELD_OFFSET

TMP_MAX

TYPE_ATTRIBUTE_ABSTRACT

TYPE_ATTRIBUTE_ANSI_CLASS

TYPE_ATTRIBUTE_AUTO_CLASS

TYPE_ATTRIBUTE_AUTO_LAYOUT

TYPE_ATTRIBUTE_BEFORE_FIELD_INIT

TYPE_ATTRIBUTE_CLASS

TYPE_ATTRIBUTE_CLASS_SEMANTIC_MASK

TYPE_ATTRIBUTE_EXPLICIT_LAYOUT

TYPE_ATTRIBUTE_FORWARDER

TYPE_ATTRIBUTE_HAS_SECURITY

TYPE_ATTRIBUTE_IMPORT

TYPE_ATTRIBUTE_INTERFACE

TYPE_ATTRIBUTE_LAYOUT_MASK

TYPE_ATTRIBUTE_NESTED_ASSEMBLY

TYPE_ATTRIBUTE_NESTED_FAMILY

TYPE_ATTRIBUTE_NESTED_FAM_AND_ASSEM

TYPE_ATTRIBUTE_NESTED_FAM_OR_ASSEM

TYPE_ATTRIBUTE_NESTED_PRIVATE

TYPE_ATTRIBUTE_NESTED_PUBLIC

TYPE_ATTRIBUTE_NOT_PUBLIC

TYPE_ATTRIBUTE_PUBLIC

TYPE_ATTRIBUTE_RESERVED_MASK

TYPE_ATTRIBUTE_RT_SPECIAL_NAME

TYPE_ATTRIBUTE_SEALED

TYPE_ATTRIBUTE_SEQUENTIAL_LAYOUT

TYPE_ATTRIBUTE_SERIALIZABLE

TYPE_ATTRIBUTE_SPECIAL_NAME

TYPE_ATTRIBUTE_STRING_FORMAT_MASK

TYPE_ATTRIBUTE_UNICODE_CLASS

TYPE_ATTRIBUTE_VISIBILITY_MASK

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WCONTINUED

WEXITED

WINT_MAX

WINT_MIN

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

_ALLOCA_H

_ASSERT_H

_ATFILE_SOURCE

_BITS_BYTESWAP_H

_BITS_ENDIANNESS_H

_BITS_ENDIAN_H

_BITS_PTHREADTYPES_ARCH_H

_BITS_PTHREADTYPES_COMMON_H

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_STDIO_LIM_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_TYPES_LOCALE_T_H

_BITS_TYPES___LOCALE_T_H

_BITS_UINTN_IDENTITY_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_ENDIAN_H

_FEATURES_H

_IOFBF

_IOLBF

_IONBF

_IO_EOF_SEEN

_IO_ERR_SEEN

_IO_USER_LOCK

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_STDIO_H

_STDLIB_H

_STRINGS_H

_STRING_H

_STRUCT_TIMESPEC

_SYS_CDEFS_H

_SYS_SELECT_H

_SYS_TYPES_H

_THREAD_MUTEX_INTERNAL_H

_THREAD_SHARED_TYPES_H

__BIG_ENDIAN

__BIT_TYPES_DEFINED__

__BYTE_ORDER

__FD_SETSIZE

__FILE_defined

__FLOAT_WORD_ORDER

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNUC_VA_LIST

__GNU_LIBRARY__

__HAVE_DISTINCT_FLOAT16

__HAVE_DISTINCT_FLOAT32

__HAVE_DISTINCT_FLOAT32X

__HAVE_DISTINCT_FLOAT64

__HAVE_DISTINCT_FLOAT64X

__HAVE_DISTINCT_FLOAT128

__HAVE_DISTINCT_FLOAT128X

__HAVE_FLOAT16

__HAVE_FLOAT32

__HAVE_FLOAT32X

__HAVE_FLOAT64

__HAVE_FLOAT64X

__HAVE_FLOAT64X_LONG_DOUBLE

__HAVE_FLOAT128

__HAVE_FLOAT128X

__HAVE_FLOATN_NOT_TYPEDEF

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__LITTLE_ENDIAN

__OFF_T_MATCHES_OFF64_T

__PDP_ENDIAN

__PTHREAD_MUTEX_HAVE_PREV

__RLIM_T_MATCHES_RLIM64_T

__SIZEOF_PTHREAD_ATTR_T

__SIZEOF_PTHREAD_BARRIERATTR_T

__SIZEOF_PTHREAD_BARRIER_T

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WALL

__WCLONE

__WCOREFLAG

__WNOTHREAD

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__W_CONTINUED

____FILE_defined

_____fpos64_t_defined

_____fpos_t_defined

____mbstate_t_defined

__clock_t_defined

__clockid_t_defined

__glibc_c99_flexarr_available

__have_pthread_attr_t

__ldiv_t_defined

__lldiv_t_defined

__sigset_t_defined

__struct_FILE_defined

__time_t_defined

__timer_t_defined

__timeval_defined

ipv6AddressSize

kCustomAttributeIndexInvalid

kDefaultValueIndexNull

kEventIndexInvalid

kFieldIndexInvalid

kGenericContainerIndexInvalid

kGenericParameterIndexInvalid

kIl2CppInt32Max

kIl2CppInt32Min

kIl2CppInt64Max

kIl2CppInt64Min

kIl2CppIntPtrMax

kIl2CppIntPtrMin

kIl2CppOffsetOfArrayBounds

kIl2CppOffsetOfArrayLength

kIl2CppSizeOfArray

kIl2CppUInt32Max

kInteropDataIndexInvalid

kInvalidIl2CppMethodSlot

kMethodIndexInvalid

kPropertyIndexInvalid

kPublicKeyByteLength

kRGCTXIndexInvalid

kStringLiteralIndexInvalid

kTypeDefinitionIndexInvalid

kTypeIndexInvalid

Statics

il2cpp_defaults

kIl2CppNewLine

kIl2CppUInt64Max

kIl2CppUIntPtrMax

stderr

stdin

stdout

Traits

WrapRaw

Safe wrapper around a raw il2cpp type which can be used in its place

Functions

assembly_get_image⚠

class_from_il2cpp_type⚠

class_from_name⚠

class_get_method_from_name⚠

class_is_assignable_from⚠

domain_get⚠

domain_get_assemblies⚠

field_get_value⚠

field_set_value⚠

format_exception⚠

runtime_invoke⚠

string_new_len⚠

type_get_name⚠

Type Definitions

AssemblyIndex

CreateCCWFunc

CustomAttributeIndex

CustomAttributesCacheGenerator

DefaultValueDataIndex

DefaultValueIndex

EncodedMethodIndex

EventIndex

FILE

FieldIndex

GenericContainerIndex

GenericInstIndex

GenericMethodIndex

GenericParameterConstraintIndex

GenericParameterIndex

IL2CPP_VARIANT_BOOL

Il2CppBacktraceFunc

Il2CppCallConvention

Il2CppCallType

Il2CppChar

Il2CppCharSet

Il2CppDecimalCompareResult

Il2CppFrameWalkFunc

Il2CppGCEvent

Il2CppHString

Il2CppLogCallback

Il2CppMetadataUsage

Il2CppMethodPointer

Il2CppMethodSlot

Il2CppNativeChar

Il2CppProfileAllocFunc

Il2CppProfileFileIOFunc

Il2CppProfileFileIOKind

Il2CppProfileFlags

Il2CppProfileFunc

Il2CppProfileGCFunc

Il2CppProfileGCResizeFunc

Il2CppProfileMethodFunc

Il2CppProfileThreadFunc

Il2CppRGCTXDataType

Il2CppResourceLocation

Il2CppRuntimeUnhandledExceptionPolicy

Il2CppSetFindPlugInCallback

Il2CppStat

Il2CppTokenType

Il2CppTypeEnum

Il2CppTypeNameFormat

Il2CppVTable

Il2CppVarType

Il2CppWindowsRuntimeTypeKind

ImageIndex

InterfaceOffsetIndex

InterfacesIndex

InteropDataIndex

InvokerMethod

MethodIndex

MethodVariableKind

NestedTypeIndex

PInvokeMarshalCleanupFunc

PInvokeMarshalFromNativeFunc

PInvokeMarshalToNativeFunc

ParameterIndex

PropertyIndex

RGCTXIndex

SequencePointKind

StringIndex

StringLiteralIndex

SynchronizationContextCallback

TypeDefinitionIndex

TypeIndex

VTableIndex

_Float32

_Float32x

_Float64

_Float64x

_IO_lock_t

__FILE

__blkcnt64_t

__blkcnt_t

__blksize_t

__builtin_va_list

__caddr_t

__clock_t

__clockid_t

__compar_fn_t

__daddr_t

__dev_t

__fd_mask

__fpos64_t

__fpos_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__gnuc_va_list

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__locale_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__pthread_list_t

__pthread_slist_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__thrd_t

__time_t

__timer_t

__tss_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

blkcnt_t

blksize_t

caddr_t

clock_t

clockid_t

daddr_t

dev_t

fd_mask

fpos_t

fsblkcnt_t

fsfilcnt_t

fsid_t

gid_t

id_t

il2cpp_WorldChangedCallback

il2cpp_array_lower_bound_t

il2cpp_array_size_t

il2cpp_hresult_t

il2cpp_register_object_callback

ino_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

key_t

locale_t

loff_t

mode_t

nlink_t

off_t

pid_t

pthread_key_t

pthread_once_t

pthread_spinlock_t

pthread_t

quad_t

register_t

sigset_t

size_t

ssize_t

suseconds_t

time_t

timer_t

u_char

u_int

u_int8_t

u_int16_t

u_int32_t

u_int64_t

u_long

u_quad_t

u_short

uid_t

uint

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

ulong

ushort

va_list

wchar_t

Unions

Il2CppDecimal__bindgen_ty_1

Il2CppDecimal__bindgen_ty_2

Il2CppDouble_double

Il2CppHStringHeader__bindgen_ty_1

Il2CppObject__bindgen_ty_1

Il2CppRGCTXData

Il2CppRGCTXDefinitionData

Il2CppSingle_float

Il2CppType__bindgen_ty_1

Il2CppVariant__bindgen_ty_1

Il2CppVariant__bindgen_ty_1___tagVARIANT__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_2

MethodInfo__bindgen_ty_1

MethodInfo__bindgen_ty_2

__mbstate_t__bindgen_ty_1

__pthread_cond_s__bindgen_ty_1

__pthread_cond_s__bindgen_ty_2

pthread_attr_t

pthread_barrier_t

pthread_barrierattr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t