libil2cpp

Struct Il2CppClass

pub struct Il2CppClass(/* private fields */);

An il2cpp class

Implementations

impl Il2CppClass

pub fn find(namespace: &str, name: &str) -> Option<&'static Self>

Find a class by namespace and name

pub fn find_generic<G>(namespace: &str, name: &str) -> Option<&'static Self>
where G: Generics,

Finds a generic method by namespace, name and generic parameters

pub fn find_method<A, R, const N: usize>( &self, name: &str, ) -> Result<&'static MethodInfo, FindMethodError>
where A: Arguments<N>, R: Returned,

Find a method belonging to the class or its parents by name with type checking

pub fn find_static_method<A, R, const N: usize>( &self, name: &str, ) -> Result<&'static MethodInfo, FindMethodError>
where A: Arguments<N>, R: Returned,

Find a static method belonging to the class by name with type checking

pub fn find_method_callee<T, P, R>( &self, name: &str, ) -> Result<&'static MethodInfo, FindMethodError>
where T: ThisParameter, P: Parameters, R: Return,

Find a method belonging to the class or its parents by name with type checking from a callee perspective

pub fn find_method_unchecked( &self, name: &str, parameters_count: usize, ) -> Result<&'static MethodInfo, FindMethodError>

Find a method belonging to the class or its parents by name and parameter count, without type checking

pub fn find_field(&self, name: &str) -> Option<&FieldInfo>

Find a field belonging to the class or its parents by name

pub fn make_generic<G>( &self, ) -> Result<Option<&'static Self>, Gc<Il2CppException>>
where G: Generics,

Instanciates a generic class template with the provided generic arguments

pub fn instantiate<T>(&self) -> &'static mut T
where for<'a> T: Type<Held<'a> = Option<&'a mut T>>,

Instanciates an object of the class

pub fn invoke<A, R, const N: usize>(&self, name: &str, args: A) -> Result<R>
where A: Arguments<N>, R: Returned,

Invokes the static method with the given name using the given arguments, with type checking

pub fn invoke_void<A, const N: usize>(&self, name: &str, args: A) -> Result<()>
where A: Arguments<N>,

Invokes the static void method with the given name using the given arguments, with type checking

pub fn name(&self) -> Cow<'_, str>

Name of the class

pub fn namespace(&self) -> Cow<'_, str>

Namespace containing the class

pub fn methods(&self) -> &[&'static MethodInfo]

Methods of the class

pub fn fields(&self) -> &'_ [FieldInfo]

Fields of the class

pub fn parent(&self) -> Option<&Self>

Parent of the class, if it inherits from any

pub fn hierarchy(&self) -> Hierarchy<'_>

Iterator over the class hierarchy, starting with the class itself

pub fn implemented_interfaces(&self) -> &[&Self]

Interfaces this class implements

pub fn nested_types(&self) -> &[&Self]

Nested types of the class

pub fn is_assignable_from(&self, other: &Self) -> bool

Whether the class is assignable from other

pub fn this_arg_ty(&self) -> &Il2CppType

Il2CppType of this for the class

pub fn byval_arg_ty(&self) -> &Il2CppType

Il2CppType of byval arguments for the class

pub fn ty(&self) -> &Il2CppType

Il2CppType of the class