Struct libil2cpp::Il2CppString

#[repr(transparent)]
pub struct Il2CppString(_);

An il2cpp string

Implementations

impl Il2CppString

pub fn new(s: impl AsRef<str>) -> &'static mut Self

Creates a new string from a Rust string

pub fn to_string(&self) -> Result<String, FromUtf16Error>

Converts the string to a Rust string, returning an error if its utf-16 data is invalid

pub fn to_string_lossy(&self) -> String

Converts the string to a Rust string, replacing any of its invalid data with the replacement character (U+FFFD)

pub fn as_utf16(&self) -> &[u16]

Returns the string’s utf-16 data as a slice

pub fn as_utf16_mut(&mut self) -> &mut [u16]

Returns the string’s utf-16 data as a mutable slice