Struct libil2cpp::MethodInfo

#[repr(transparent)]
pub struct MethodInfo(_);

Information about a C# method

Implementations

impl MethodInfo

pub fn invoke<T, A, R, const N: usize>(
&self,
this: T,
args: A
) -> Result<R, &mut Il2CppException> where
T: ThisArgument,
A: Arguments<N>,
R: Returned,

Invoke this method, type checking against its signature with the provided instance, arguments and return type

pub unsafe fn invoke_unchecked<T, A, R, const N: usize>(
&self,
this: T,
args: A
) -> Result<R, &mut Il2CppException> where
T: ThisArgument,
A: Arguments<N>,
R: Returned,

Invoke this method with the given instance and arguments and converting the result to the specified type, without type checking

Safety

To be safe, the provided types have to match the method signature

pub fn reflection_object(&self) -> &Il2CppReflectionMethod

Il2CppReflectionMethod which represents the method

pub fn name(&self) -> Cow<'_, str>

Name of the method

pub fn class(&self) -> &Il2CppClass

Class the method is from

pub fn return_ty(&self) -> &Il2CppType

Return type of the method

pub fn parameters(&self) -> &'_ [ParameterInfo]

Parameters the method takes

pub fn is_static(&self) -> bool

Whether the method is static

pub fn is_abstract(&self) -> bool

Whether the method is abstract

pub fn is_virtual(&self) -> bool

Whether the method is virtual

pub fn is_generic(&self) -> bool

Whether the method is generic