Trait libil2cpp::Argument

pub unsafe trait Argument {
    type Type: Any;

    fn matches(ty: &Il2CppType) -> bool;
    fn invokable(&mut self) -> *mut c_void;
}

Trait implemented by types that can be used as C# method arguments

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Required Associated Types

type Type: Any

Normalized type of the argument, useful for caching

Required Methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# argument with the given Il2CppType to call a method

fn invokable(&mut self) -> *mut c_void

Returns an untyped pointer which can be used as a libil2cpp argument