Trait libil2cpp::Argument

pub unsafe trait Argument {
    type Type: Any;
    fn matches(ty: &Il2CppType) -> bool;

    fn invokable(&self) -> *mut c_void;
}

Trait implemented by types that can be used as C# method arguments

Safety

Interfaces depending on this trait assume that all of its methods are correct in an il2cpp context

Associated Types

type Type: Any

Normalized type of the argument, useful for caching

Required methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# argument with the given Il2CppType to call a method

fn invokable(&self) -> *mut c_void

Returns an untyped pointer which can be used as an argument to invoke C# methods