Trait libil2cpp::Arguments

pub unsafe trait Arguments<const N: usize> {
    type Type: Any;
    fn matches(method: &MethodInfo) -> bool;

    fn invokable(&mut self) -> [*mut c_void; N];
}

Trait implemented by types that can be used as a collection of C# method arguments

Note

You should most likely not be implementing this trait yourself

Safety

The implementation must be correct

Associated Types

type Type: Any

Normalized type of the arguments, useful for caching

Required methods

fn matches(method: &MethodInfo) -> bool

Checks whether the type can be used as a C# argument collection for the given MethodInfo

fn invokable(&mut self) -> [*mut c_void; N]

Returns an array of untyped pointer which can be used to invoke C# methods