Trait libil2cpp::Parameters

pub unsafe trait Parameters {
    const COUNT: usize;

    fn matches(method: &MethodInfo) -> bool;
}

Trait implemented by types that can be used as a collection of C# method parameters

Note

You should most likely not be implementing this trait yourself

Safety

The implementation must be correct

Required Associated Constants

const COUNT: usize

Parameter count

Required Methods

fn matches(method: &MethodInfo) -> bool

Checks whether the type can be used as a C# parameter collection for the given MethodInfo