Trait libil2cpp::Return

pub unsafe trait Return {
    type Actual;
    type Type: Any;
    fn matches(ty: &Il2CppType) -> bool;

    fn into_actual(self) -> Self::Actual;

    fn from_actual(actual: Self::Actual) -> Self;
}

Trait implemented by types that can be used as return types for C# methods

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Associated Types

type Actual

Type of the actual return value

type Type: Any

Normalized type of the return value, useful for caching

Required methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# return type of the given Il2CppType

fn into_actual(self) -> Self::Actual

Converts from the desired type into the actual one

fn from_actual(actual: Self::Actual) -> Self

Converts from the actual type to the desired one