Trait libil2cpp::Returned

pub unsafe trait Returned {
    type Type: Any;
    fn matches(ty: &Il2CppType) -> bool;

    fn from_object(object: Option<&mut Il2CppObject>) -> Self;
}

Trait implemented by types that can be used as return types from C# methods

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Associated Types

type Type: Any

Normalized type of the return type, useful for caching

Required methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# return type of the given Il2CppType

fn from_object(object: Option<&mut Il2CppObject>) -> Self

Converts the Il2CppObject returned by runtime_invoke into self