Trait libil2cpp::ThisArgument

pub unsafe trait ThisArgument {
    type Type: Any;
    fn matches(method: &MethodInfo) -> bool;

    fn invokable(&mut self) -> *mut c_void;
}

Trait implemented by types that can be used as a C# this arguments

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Associated Types

type Type: Any

Normalized type of this, useful for caching

Required methods

fn matches(method: &MethodInfo) -> bool

Checks whether the type can be used as a C# this argument for the given MethodInfo

fn invokable(&mut self) -> *mut c_void

Returns an untyped pointer which can be used as a libill2cpp this argument