Trait libil2cpp::ThisParameter

pub unsafe trait ThisParameter {
    type Actual;

    fn matches(method: &MethodInfo) -> bool;
    fn from_actual(actual: Self::Actual) -> Self;
    fn into_actual(self) -> Self::Actual;
}

Trait implemented by types that can be used as C# this method parameters

Note

You should most likely not be implementing this trait yourself, but rather the Type trait

Safety

The implementation must be correct

Required Associated Types

type Actual

Type of the actual this parameter

Required Methods

fn matches(method: &MethodInfo) -> bool

Checks whether the type can be used as a C# instance parameter for the given MethodInfo

fn from_actual(actual: Self::Actual) -> Self

Converts from the actual type to the desired one

fn into_actual(self) -> Self::Actual

Converts from the desired type into the actual one