Trait libil2cpp::Type

pub unsafe trait Type: Any {
    type Semantics: Semantics;

    const NAMESPACE: &'static str;
    const CLASS_NAME: &'static str;

    fn matches_this_argument(method: &MethodInfo) -> bool;

    fn matches_this_parameter(method: &MethodInfo) -> bool;

    fn matches_reference_argument(ty: &Il2CppType) -> bool;

    fn matches_value_argument(ty: &Il2CppType) -> bool;

    fn matches_reference_parameter(ty: &Il2CppType) -> bool;

    fn matches_value_parameter(ty: &Il2CppType) -> bool;

    fn matches_reference_returned(ty: &Il2CppType) -> bool;

    fn matches_value_returned(ty: &Il2CppType) -> bool;

    fn matches_reference_return(ty: &Il2CppType) -> bool;

    fn matches_value_return(ty: &Il2CppType) -> bool;

    fn class() -> &'static Il2CppClass { ... }
}

Trait implemented by Rust types that are also C# types

Safety

The Rust and C# types must be ABI-compatible and the trait implementation must be correct

Associated Types

type Semantics: Semantics

Semantics of the type, either Reference or Value

Associated Constants

const NAMESPACE: &'static str

Namespace containingthe class the type represents

const CLASS_NAME: &'static str

Name of the class the type represents

Required methods

fn matches_this_argument(method: &MethodInfo) -> bool

Whether the type can be used as a this argument for the given MethodInfo

fn matches_this_parameter(method: &MethodInfo) -> bool

Whether the type can be used as a this parameter for the given MethodInfo

fn matches_reference_argument(ty: &Il2CppType) -> bool

Whether a reference to the type can be used as an argument of the given Il2CppType

fn matches_value_argument(ty: &Il2CppType) -> bool

Whether a value of the type can be used as an argument of the given Il2CppType

fn matches_reference_parameter(ty: &Il2CppType) -> bool

Whether a reference to the type can be used as a parameter of the given Il2CppType

fn matches_value_parameter(ty: &Il2CppType) -> bool

Whether a value of the type can be used as a parameter of the given Il2CppType

fn matches_reference_returned(ty: &Il2CppType) -> bool

Whether a reference to the type can be used as the value of the given Il2CppType returned from a C# method

fn matches_value_returned(ty: &Il2CppType) -> bool

Whether a value of the type can be used as the value of the given Il2CppType returned from a C# method

fn matches_reference_return(ty: &Il2CppType) -> bool

Whether a reference to the type can be used as the return value of the given Il2CppType for a C# method

fn matches_value_return(ty: &Il2CppType) -> bool

Whether a value of the type can be used as the return value of the given Il2CppType for a C# method

Provided methods

fn class() -> &'static Il2CppClass

Il2CppClass of the type